Refactor MCP server into modular components
