+++ pointblank/mcp/server.py
@@ -1,15 +1,12 @@
+"""Pointblank MCP Server — tool and prompt definitions."""
+
 import json
-import logging
-import math
-import os
 import platform
 import sys
 import uuid
-import webbrowser
 from contextlib import asynccontextmanager
 from dataclasses import dataclass, field
 from datetime import datetime
-from pathlib import Path
 from typing import (
     Annotated,
     Any,
@@ -20,164 +17,42 @@ from typing import (
 )
 
 from fastmcp import Context, FastMCP
-from fastmcp.prompts.prompt import Message
 
 import pointblank as pb
-
-# Detect if we're running in a test environment
-TESTING_MODE = (
-    "pytest" in sys.modules
-    or os.environ.get("PYTEST_CURRENT_TEST") is not None
-    or os.environ.get("POINTBLANK_TESTING") == "true"
+from pointblank.mcp._config import HAS_PANDAS, HAS_POLARS, logger
+from pointblank.mcp._prompts import (
+    prompt_add_validation_step_example,
+    prompt_create_validator,
+    prompt_get_validation_step_output,
+    prompt_interrogate_validator,
+    prompt_load_dataframe,
+)
+from pointblank.mcp._templates import AVAILABLE_TEMPLATES, get_validation_template
+from pointblank.mcp._utils import (
+    SUPPORTED_VALIDATION_TYPES,
+    clean_for_json_serialization,
+    generate_python_code_for_validator,
+    generate_validation_report_html,
+    get_available_backends,
+    get_supported_validations,
+    load_dataframe_from_path,
+    open_browser_conditionally,
+    save_dataframe_to_csv,
+    save_html_and_open,
+    validate_output_path,
+    validate_resource_id,
 )
 
-# Try to import Pandas, but make it optional
-try:
-    import pandas as pd
-
-    HAS_PANDAS = True
-except ImportError:
-    pd = None
-    HAS_PANDAS = False
-
-# Try to import other DataFrame libraries
-try:
-    import polars as pl
-
-    HAS_POLARS = True
-except ImportError:
-    pl = None
-    HAS_POLARS = False
-
-
-# Configure logging
-def _get_log_path() -> Path:
-    """Get a writable log file path, falling back to temp directory."""
-    candidates = [
-        Path.home() / ".pointblank" / "mcp_server.log",
-        Path("/tmp") / "pointblank_mcp_server.log",
-    ]
-    for p in candidates:
-        try:
-            p.parent.mkdir(parents=True, exist_ok=True)
-            return p
-        except OSError:
-            continue
-    # Last resort: no file logging
-    return None
-
-
-_log_path = _get_log_path()
-_log_handlers: list = [logging.StreamHandler()]
-if _log_path:
-    _log_handlers.append(logging.FileHandler(str(_log_path)))
-
-logging.basicConfig(
-    level=logging.INFO,
-    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-    handlers=_log_handlers,
-)
-logger = logging.getLogger(__name__)
 logger.info(f"MCP Server starting at {datetime.now()}")
 logger.info(f"Available DataFrame backends: pandas={HAS_PANDAS}, polars={HAS_POLARS}")
-logger.info("Core Pointblank visualization features available")
-
-# Maximum file size for loading (500 MB)
-_MAX_FILE_SIZE_BYTES = 500 * 1024 * 1024
-
-# Allowed file extensions for data loading
-_ALLOWED_DATA_EXTENSIONS = {".csv", ".xls", ".xlsx", ".parquet", ".json", ".jsonl"}
 
 
-def _validate_input_path(input_path: str) -> Path:
-    """
-    Validate and resolve an input file path, preventing path traversal attacks.
-
-    Ensures the path:
-    - Is resolved to an absolute path
-    - Does not contain traversal sequences
-    - Has an allowed file extension
-    - Does not exceed the maximum file size
-    """
-    p = Path(input_path).resolve()
-
-    # Block obvious traversal patterns in the raw input
-    if ".." in Path(input_path).parts:
-        raise ValueError("Path traversal ('..') is not allowed in file paths.")
-
-    # Validate extension
-    if p.suffix.lower() not in _ALLOWED_DATA_EXTENSIONS:
-        raise ValueError(
-            f"File type '{p.suffix}' is not allowed. "
-            f"Supported: {', '.join(sorted(_ALLOWED_DATA_EXTENSIONS))}"
-        )
-
-    # Check existence
-    if not p.exists():
-        raise FileNotFoundError(f"Input file not found: {p}")
-
-    if not p.is_file():
-        raise ValueError(f"Path is not a regular file: {p}")
-
-    # Check file size
-    file_size = p.stat().st_size
-    if file_size > _MAX_FILE_SIZE_BYTES:
-        raise ValueError(
-            f"File size ({file_size / 1024 / 1024:.1f} MB) exceeds the "
-            f"{_MAX_FILE_SIZE_BYTES / 1024 / 1024:.0f} MB limit."
-        )
-
-    return p
+# --- Lifespan Context ---
 
 
-def _validate_output_path(output_path: str, allowed_extensions: set[str]) -> Path:
-    """
-    Validate an output file path, preventing writes to dangerous locations.
-
-    Ensures the path:
-    - Is resolved to an absolute path
-    - Does not contain traversal sequences
-    - Has an allowed file extension
-    - Parent directory exists or can be created
-    """
-    p = Path(output_path).resolve()
-
-    # Block traversal
-    if ".." in Path(output_path).parts:
-        raise ValueError("Path traversal ('..') is not allowed in file paths.")
-
-    # Validate extension
-    if p.suffix.lower() not in allowed_extensions:
-        raise ValueError(
-            f"File type '{p.suffix}' is not allowed. "
-            f"Supported: {', '.join(sorted(allowed_extensions))}"
-        )
-
-    # Don't allow writing into system directories
-    _blocked_prefixes = ("/etc", "/usr", "/bin", "/sbin", "/var", "/sys", "/proc")
-    for prefix in _blocked_prefixes:
-        if str(p).startswith(prefix):
-            raise ValueError(f"Writing to system directory '{prefix}' is not allowed.")
-
-    # Ensure parent directory exists
-    p.parent.mkdir(parents=True, exist_ok=True)
-
-    return p
-
-
-# Type alias for DataFrame: can be Pandas or Polars or other
-if HAS_PANDAS:
-    DataFrameType = pd.DataFrame
-else:
-    DataFrameType = Any  # Fallback to Any if pandas not available
-
-
-# --- Lifespan Context: manage DataFrames and Validators ---
 @dataclass
 class AppContext:
-    # Stores loaded DataFrames: {df_id: DataFrame}
     loaded_dataframes: Dict[str, Any] = field(default_factory=dict)
-    # Stores active Pointblank Validators: {validator_id: Validate}
     active_validators: Dict[str, pb.Validate] = field(default_factory=dict)
 
 
@@ -189,16 +64,15 @@ async def app_lifespan(server: FastMCP) -> AsyncIterator[AppContext]:
     context.active_validators.clear()
 
 
-# Create FastMCP instance with version-aware dependency handling
+# --- FastMCP Instance ---
+
+
 def _create_fastmcp_instance():
     """Create FastMCP instance with backwards compatibility for dependencies parameter."""
     try:
-        # Try to get FastMCP version to determine if the `dependencies=` parameter is available
         import fastmcp
 
         version_str = getattr(fastmcp, "__version__", "0.0.0")
-
-        # Parse version and assume format like "2.11.4"
         version_parts = version_str.split(".")
         if len(version_parts) >= 3:
             major, minor, patch = (
@@ -206,341 +80,65 @@ def _create_fastmcp_instance():
                 int(version_parts[1]),
                 int(version_parts[2]),
             )
-
-            # For versions >=2.11.4, the `dependencies=` parameter is deprecated
             if (
                 (major > 2)
                 or (major == 2 and minor > 11)
                 or (major == 2 and minor == 11 and patch >= 4)
             ):
-                return FastMCP(
-                    "PointblankMCP",
-                    lifespan=app_lifespan,
-                )
+                return FastMCP("PointblankMCP", lifespan=app_lifespan)
 
-        # For older versions, use the `dependencies=` parameter
         return FastMCP(
-            "PointblankMCP",
-            lifespan=app_lifespan,
-            dependencies=["pointblank", "openpyxl"],
+            "PointblankMCP", lifespan=app_lifespan, dependencies=["pointblank", "openpyxl"]
         )
 
     except Exception:
-        # Fallback: try without dependencies first (newer approach)
-        # If that fails, try with dependencies (older approach)
         try:
-            return FastMCP(
-                "PointblankMCP",
-                lifespan=app_lifespan,
-            )
+            return FastMCP("PointblankMCP", lifespan=app_lifespan)
         except Exception:
             return FastMCP(
-                "PointblankMCP",
-                lifespan=app_lifespan,
-                dependencies=["pointblank", "openpyxl"],
+                "PointblankMCP", lifespan=app_lifespan, dependencies=["pointblank", "openpyxl"]
             )
 
 
 mcp = _create_fastmcp_instance()
 
 
-def _get_available_backends():
-    """Get list of available DataFrame backends."""
-    backends = []
-    if HAS_PANDAS:
-        backends.append("pandas")
-    if HAS_POLARS:
-        backends.append("polars")
-    return backends
+# --- Register Prompts ---
+
+mcp.prompt(
+    name="prompt_load_dataframe",
+    description="Prompt to load a DataFrame from a file into the server's context for validation.",
+    tags={"Data Management"},
+)(prompt_load_dataframe)
+
+mcp.prompt(
+    name="prompt_create_validator",
+    description="Prompt to create a Pointblank Validator for a loaded DataFrame.",
+    tags={"Validation"},
+)(prompt_create_validator)
+
+mcp.prompt(
+    name="prompt_add_validation_step_example",
+    description="Prompt to add a validation step to a Pointblank Validator.",
+    tags={"Validation"},
+)(prompt_add_validation_step_example)
+
+mcp.prompt(
+    name="prompt_get_validation_step_output",
+    description="Prompt to get validation output by specifying either a step index or a sundered type.",
+    tags={"Validation"},
+)(prompt_get_validation_step_output)
+
+mcp.prompt(
+    name="prompt_interrogate_validator",
+    description="Prompt to run validations and generate reports with Python code.",
+    tags={"Validation"},
+)(prompt_interrogate_validator)
 
 
-def _save_dataframe_to_csv(df: Any, output_path: Path) -> None:
-    """Save DataFrame to CSV in a backend-agnostic way."""
-    if HAS_PANDAS and hasattr(df, "to_csv") and hasattr(df, "index"):
-        # Pandas DataFrame
-        df.to_csv(output_path, index=False)
-    elif HAS_POLARS and hasattr(df, "write_csv"):
-        # Polars DataFrame
-        df.write_csv(output_path)
-    else:
-        # Fallback: try to convert to pandas if available
-        if HAS_PANDAS:
-            if hasattr(df, "to_pandas"):
-                # Polars to pandas conversion
-                df.to_pandas().to_csv(output_path, index=False)
-            else:
-                # Try direct pandas constructor
-                pd.DataFrame(df).to_csv(output_path, index=False)
-        else:
-            raise TypeError(f"Unsupported DataFrame type '{type(df).__name__}' for CSV export.")
-
-
-def _open_browser_conditionally(url: str) -> None:
-    """Open browser only if not in testing mode."""
-    if not TESTING_MODE:
-        webbrowser.open(url)
-    else:
-        logger.debug(f"Browser opening suppressed in testing mode for: {url}")
-
-
-def _save_html_and_open(html_content: str, title: str, filename_prefix: str) -> str:
-    """
-    Wrap HTML content in a styled document, save to a file, and open in browser.
-
-    Returns a user-facing message about the result.
-    """
-    full_html = f"""<!DOCTYPE html>
-<html>
-<head>
-    <title>{title}</title>
-    <meta charset="utf-8">
-    <style>
-        body {{
-            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-            margin: 20px;
-            background-color: #f8f9fa;
-        }}
-        .table-container {{
-            background: white;
-            padding: 20px;
-            border-radius: 8px;
-            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-            overflow-x: auto;
-        }}
-    </style>
-</head>
-<body>
-    <div class="table-container">
-        {html_content}
-    </div>
-</body>
-</html>
-"""
-    html_filename = f"{filename_prefix}_{datetime.now().strftime('%Y%m%d_%H%M%S')}.html"
-    html_path = Path.cwd() / html_filename
-
-    if TESTING_MODE:
-        return (
-            f"HTML generated (file creation skipped during testing)\n\nFile location: {html_path}"
-        )
-
-    try:
-        with open(html_path, "w", encoding="utf-8") as f:
-            f.write(full_html)
-        _open_browser_conditionally(f"file://{html_path}")
-        return f"HTML saved and opened in default browser!\n\nFile location: {html_path}"
-    except Exception as e:
-        return f"Error saving HTML file: {str(e)}"
-
-
-def _load_dataframe_from_path(input_path: str, backend: str = "auto") -> Any:
-    """Load DataFrame from file using specified backend or auto-detect."""
-    p_path = _validate_input_path(input_path)
-
-    # Auto-detect backend
-    if backend == "auto":
-        if HAS_PANDAS:
-            backend = "pandas"
-        elif HAS_POLARS:
-            backend = "polars"
-        else:
-            raise ImportError("No DataFrame library available. Install pandas or polars.")
-
-    # Load with specified backend
-    if backend == "pandas":
-        if not HAS_PANDAS:
-            raise ImportError("Pandas not available. Install with: pip install pandas")
-
-        if p_path.suffix.lower() == ".csv":
-            return pd.read_csv(p_path)
-        elif p_path.suffix.lower() in [".xls", ".xlsx"]:
-            return pd.read_excel(p_path, engine="openpyxl")
-        elif p_path.suffix.lower() == ".parquet":
-            return pd.read_parquet(p_path)
-        elif p_path.suffix.lower() == ".json":
-            return pd.read_json(p_path)
-        elif p_path.suffix.lower() == ".jsonl":
-            return pd.read_json(p_path, lines=True)
-    elif backend == "polars":
-        if not HAS_POLARS:
-            raise ImportError("Polars not available. Install with: pip install polars")
-
-        if p_path.suffix.lower() == ".csv":
-            return pl.read_csv(p_path)
-        elif p_path.suffix.lower() == ".parquet":
-            return pl.read_parquet(p_path)
-        elif p_path.suffix.lower() == ".json":
-            return pl.read_json(p_path)
-        elif p_path.suffix.lower() == ".jsonl":
-            return pl.read_ndjson(p_path)
-        elif p_path.suffix.lower() in [".xls", ".xlsx"]:
-            # Polars doesn't directly support Excel, fall back to pandas if available
-            if HAS_PANDAS:
-                return pd.read_excel(p_path, engine="openpyxl")
-            else:
-                raise ValueError(
-                    "Excel files require pandas. Install with: pip install pandas openpyxl"
-                )
-    else:
-        raise ValueError(f"Unsupported backend: {backend}. Available: {_get_available_backends()}")
-
-    raise ValueError(
-        f"Unsupported file type: {p_path.suffix}. Please use CSV, Excel, Parquet, JSON, or JSONL."
-    )
-
-
-def _generate_python_code_for_validator(
-    validator: pb.Validate, validator_id: str, df_path: Optional[str] = None
-) -> str:
-    """
-    Generate Python code equivalent for reproducing the validation using fluent interface.
-    """
-    # Start building the Python code
-    code_lines = [
-        "# Generated Python code for Pointblank validation",
-        "import pointblank as pb",
-        "",
-        "# Load your data",
-    ]
-
-    if df_path:
-        code_lines.extend(
-            [
-                f"# Original file: {df_path}",
-                f"df = pb.load_dataset('{df_path}')  # Adjust path as needed",
-            ]
-        )
-    else:
-        code_lines.extend(
-            [
-                "# Replace 'your_data.csv' with your actual data file",
-                "df = pb.load_dataset('your_data.csv')",
-            ]
-        )
-
-    # Get validation steps from the validator's method chain
-    validation_methods = []
-
-    try:
-        # Access the validator's validation steps
-        # Reconstruct based on the report data
-        json_report = validator.get_json_report()
-        validation_data = json.loads(json_report)
-
-        for step in validation_data:
-            assertion_type = step.get("assertion_type", "")
-            column = step.get("column", "")
-            values = step.get("values", None)
-            inclusive = step.get("inclusive", None)
-
-            if assertion_type == "rows_distinct":
-                validation_methods.append("    .rows_distinct()")
-
-            elif assertion_type == "col_vals_not_null":
-                if column:
-                    validation_methods.append(f"    .col_vals_not_null(columns='{column}')")
-
-            elif assertion_type == "col_vals_between":
-                if column and values and len(values) >= 2:
-                    left, right = values[0], values[1]
-                    validation_methods.append(
-                        f"    .col_vals_between(columns='{column}', left={left}, right={right})"
-                    )
-
-            elif assertion_type == "col_vals_ge":
-                if column and values is not None:
-                    validation_methods.append(
-                        f"    .col_vals_ge(columns='{column}', value={values})"
-                    )
-
-            elif assertion_type == "col_vals_gt":
-                if column and values is not None:
-                    validation_methods.append(
-                        f"    .col_vals_gt(columns='{column}', value={values})"
-                    )
-
-            elif assertion_type == "col_vals_le":
-                if column and values is not None:
-                    validation_methods.append(
-                        f"    .col_vals_le(columns='{column}', value={values})"
-                    )
-
-            elif assertion_type == "col_vals_lt":
-                if column and values is not None:
-                    validation_methods.append(
-                        f"    .col_vals_lt(columns='{column}', value={values})"
-                    )
-
-            elif assertion_type == "col_vals_in_set":
-                if column and values:
-                    set_values = repr(values) if isinstance(values, list) else f"[{repr(values)}]"
-                    validation_methods.append(
-                        f"    .col_vals_in_set(columns='{column}', set={set_values})"
-                    )
-
-            elif assertion_type == "col_vals_regex":
-                if column and values:
-                    pattern = values if isinstance(values, str) else str(values)
-                    validation_methods.append(
-                        f"    .col_vals_regex(columns='{column}', pattern=r'{pattern}')"
-                    )
-
-            elif assertion_type == "col_exists":
-                if column:
-                    if isinstance(column, list):
-                        cols = repr(column)
-                    else:
-                        cols = f"'{column}'"
-                    validation_methods.append(f"    .col_exists(columns={cols})")
-
-    except Exception as e:
-        validation_methods.append(f"    # Error reconstructing validation steps: {e}")
-        validation_methods.append("    # Please manually add your validation steps")
-
-    # Build the fluent interface chain
-    code_lines.extend(
-        [
-            "",
-            "# Create validator, add validation steps, and interrogate",
-            "validator = (",
-            "    pb.Validate(df)",
-        ]
-    )
-
-    # Add all validation methods
-    code_lines.extend(validation_methods)
-
-    # Close the chain with interrogate
-    code_lines.extend(
-        ["    .interrogate()", ")", "", "# View HTML report", "validator.get_tabular_report()"]
-    )
-
-    return "\n".join(code_lines)
-
-
-def _generate_validation_report_html(validator: pb.Validate, validator_id: str) -> str:
-    """
-    Generate an HTML report table for validation results and save to file.
-    Uses Pointblank's built-in get_tabular_report() and as_raw_html() methods.
-    Returns the file path.
-    """
-    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-    filename = f"pointblank_validation_report_{validator_id}_{timestamp}.html"
-    file_path = Path.cwd() / filename
-
-    # Skip file generation during testing
-    if TESTING_MODE:
-        logger.debug(f"Skipping HTML file generation during testing: {filename}")
-        return str(file_path.resolve())
-
-    gt_report = validator.get_tabular_report()
-    html_content = gt_report.as_raw_html()
-
-    with open(file_path, "w", encoding="utf-8") as f:
-        f.write(html_content)
-
-    logger.info(f"Validation report HTML saved to: {file_path}")
-    return str(file_path.resolve())
+# =============================================================================
+# TOOLS: Data Management
+# =============================================================================
 
 
 @dataclass
@@ -565,29 +163,22 @@ async def load_dataframe(
     ] = None,
     backend: Annotated[
         str,
-        "DataFrame backend to use: 'auto', 'pandas', or 'polars'. Default is 'auto' (uses pandas if available, then polars).",
+        "DataFrame backend to use: 'auto', 'pandas', or 'polars'. Default is 'auto'.",
     ] = "auto",
 ) -> DataFrameInfo:
-    """
-    Loads a DataFrame from the specified CSV, Excel, or Parquet file into the server's context.
-    Assigns a unique ID to the DataFrame for later reference.
-    If df_id is not provided, a new one will be generated.
-    Supports multiple DataFrame backends (pandas, polars) for flexibility.
-    Returns the DataFrame ID and basic information (shape, columns).
-    """
+    """Load a DataFrame from file into the server's context."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
 
-    # Check available backends
-    available_backends = _get_available_backends()
+    available_backends = get_available_backends()
     if not available_backends:
         raise RuntimeError("No DataFrame library available. Install pandas or polars.")
 
-    # Inform user about available backends
     await ctx.report_progress(10, 100, f"Available backends: {', '.join(available_backends)}")
 
-    df = _load_dataframe_from_path(input_path, backend)
+    df = load_dataframe_from_path(input_path, backend)
 
     effective_df_id = df_id if df_id else f"df_{uuid.uuid4().hex[:8]}"
+    effective_df_id = validate_resource_id(effective_df_id, "DataFrame")
 
     if effective_df_id in app_ctx.loaded_dataframes:
         raise ValueError(
@@ -596,7 +187,6 @@ async def load_dataframe(
 
     app_ctx.loaded_dataframes[effective_df_id] = df
 
-    # Get DataFrame info in a backend-agnostic way
     shape = df.shape
     columns = list(df.columns)
 
@@ -606,7 +196,7 @@ async def load_dataframe(
 
     return DataFrameInfo(
         df_id=effective_df_id,
-        shape=(int(shape[0]), int(shape[1])),  # Convert to Python ints
+        shape=(int(shape[0]), int(shape[1])),
         columns=columns,
     )
 
@@ -617,28 +207,22 @@ async def load_dataframe(
     tags={"Data Management"},
 )
 async def list_available_backends(ctx: Context) -> Dict[str, Any]:
-    """
-    Returns information about available DataFrame backends and their capabilities.
-    """
-    backends = _get_available_backends()
+    """Returns information about available DataFrame backends and their capabilities."""
+    backends = get_available_backends()
 
     backend_info = {}
-    for backend in backends:
-        if backend == "pandas":
+    for b in backends:
+        if b == "pandas":
             backend_info["pandas"] = {
                 "available": True,
                 "supports": ["CSV", "Excel", "Parquet", "JSON"],
-                "excel_engine": "openpyxl" if HAS_PANDAS else None,
             }
-        elif backend == "polars":
+        elif b == "polars":
             backend_info["polars"] = {
                 "available": True,
-                "supports": ["CSV", "Parquet"],
+                "supports": ["CSV", "Parquet", "JSON"],
                 "notes": "Excel requires fallback to pandas",
             }
-
-    if not backends:
-        backend_info["warning"] = "No DataFrame backends available. Install pandas or polars."
 
     return {
         "available_backends": backends,
@@ -655,9 +239,7 @@ async def list_available_backends(ctx: Context) -> Dict[str, Any]:
     tags={"Data Management"},
 )
 async def list_loaded_dataframes(ctx: Context) -> Dict[str, Any]:
-    """
-    Returns information about all DataFrames currently loaded in the server.
-    """
+    """Returns information about all DataFrames currently loaded in the server."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
 
     dataframes_info = {}
@@ -665,18 +247,14 @@ async def list_loaded_dataframes(ctx: Context) -> Dict[str, Any]:
         try:
             shape = df.shape
             columns = list(df.columns)
-            # Detect DataFrame type
             df_type = "pandas" if hasattr(df, "to_csv") and hasattr(df, "index") else "polars"
 
             dataframes_info[df_id] = {
-                "shape": [int(shape[0]), int(shape[1])],  # Convert to Python ints
+                "shape": [int(shape[0]), int(shape[1])],
                 "columns": columns,
                 "column_count": len(columns),
-                "row_count": int(shape[0]),  # Convert to Python int
+                "row_count": int(shape[0]),
                 "backend": df_type,
-                "memory_usage_mb": round(float(df.memory_usage(deep=True).sum()) / 1024 / 1024, 2)
-                if df_type == "pandas"
-                else "N/A",
             }
         except Exception as e:
             dataframes_info[df_id] = {
@@ -684,111 +262,12 @@ async def list_loaded_dataframes(ctx: Context) -> Dict[str, Any]:
                 "backend": "unknown",
             }
 
-    result = {
-        "loaded_dataframes": dataframes_info,
-        "total_count": len(dataframes_info),
-        "memory_usage_summary": {
-            "total_mb": sum(
-                info.get("memory_usage_mb", 0)
-                for info in dataframes_info.values()
-                if isinstance(info.get("memory_usage_mb"), (int, float))
-            )
-        },
-    }
-
-    # Clean the result to ensure all numpy types are converted
-    return clean_for_json_serialization(result)
-
-
-@mcp.tool(
-    name="analyze_data_quality",
-    description="Analyze data quality using Pointblank's DataScan class.",
-    tags={"Data Analysis"},
-)
-async def analyze_data_quality(
-    ctx: Context,
-    df_id: Annotated[str, "ID of the DataFrame to analyze."],
-) -> Dict[str, Any]:
-    """
-    Analyze data quality using Pointblank's built-in DataScan functionality.
-    """
-    app_ctx: AppContext = ctx.request_context.lifespan_context
-
-    if df_id not in app_ctx.loaded_dataframes:
-        raise ValueError(f"DataFrame ID '{df_id}' not found.")
-
-    df = app_ctx.loaded_dataframes[df_id]
-
-    await ctx.report_progress(20, 100, "Starting data quality analysis...")
-
-    try:
-        # Use Pointblank's DataScan class for profiling
-        scanner = pb.DataScan(data=df)
-
-        await ctx.report_progress(60, 100, "Running DataScan analysis...")
-
-        # Get the JSON representation which properly handles numpy types
-        profile_json = scanner.to_json()
-
-        await ctx.report_progress(80, 100, "Processing results...")
-
-        # Parse it back to a dictionary and clean any problematic values
-        profile_results = json.loads(profile_json)
-        cleaned_results = clean_for_json_serialization(profile_results)
-
-        await ctx.report_progress(100, 100, "Data quality analysis complete!")
-
-        return {"status": "success", "df_id": df_id, "analysis": cleaned_results}
-
-    except Exception as e:
-        error_msg = f"Error during data quality analysis: {str(e)}"
-        await ctx.report_progress(100, 100, error_msg)
-        raise ValueError(error_msg)
-
-
-@mcp.tool(
-    name="list_active_validators",
-    description="List all validators currently active in the server context.",
-    tags={"Validation"},
-)
-async def list_active_validators(ctx: Context) -> Dict[str, Any]:
-    """
-    Returns information about all active validators in the server.
-    """
-    app_ctx: AppContext = ctx.request_context.lifespan_context
-
-    validators_info = {}
-    for validator_id, validator in app_ctx.active_validators.items():
-        try:
-            # Get validator metadata
-            table_name = getattr(validator, "tbl_name", "Unknown")
-            label = getattr(validator, "label", "No label")
-
-            # Check if interrogated
-            is_interrogated = (
-                hasattr(validator, "time_processed") and validator.time_processed is not None
-            )
-
-            # Count validation steps (this is a rough estimate)
-            step_count = len(getattr(validator, "_validation_set", []))
-
-            validators_info[validator_id] = {
-                "table_name": table_name,
-                "label": label,
-                "is_interrogated": is_interrogated,
-                "validation_steps_count": step_count,
-                "last_processed": str(getattr(validator, "time_processed", "Never")),
-            }
-        except Exception as e:
-            validators_info[validator_id] = {"error": f"Failed to get info: {str(e)}"}
-
-    return {
-        "active_validators": validators_info,
-        "total_count": len(validators_info),
-        "interrogated_count": sum(
-            1 for info in validators_info.values() if info.get("is_interrogated", False)
-        ),
-    }
+    return clean_for_json_serialization(
+        {
+            "loaded_dataframes": dataframes_info,
+            "total_count": len(dataframes_info),
+        }
+    )
 
 
 @mcp.tool(
@@ -800,91 +279,59 @@ async def delete_dataframe(
     ctx: Context,
     df_id: Annotated[str, "ID of the DataFrame to delete."],
 ) -> Dict[str, str]:
-    """
-    Removes a DataFrame from the server context and frees up memory.
-    Also removes any validators that were using this DataFrame.
-    """
+    """Removes a DataFrame from the server context and frees up memory."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
+    df_id = validate_resource_id(df_id, "DataFrame")
 
     if df_id not in app_ctx.loaded_dataframes:
         raise ValueError(f"DataFrame ID '{df_id}' not found.")
 
-    # Remove the DataFrame
     del app_ctx.loaded_dataframes[df_id]
 
-    # Find and remove validators that might be using this DataFrame
-    validators_to_remove = []
-    for validator_id, validator in app_ctx.active_validators.items():
-        # This is a heuristic as we can't easily determine which DataFrame a validator uses
-        # In a more sophisticated implementation, we'd track this relationship
-        try:
-            if hasattr(validator, "tbl_name") and df_id in validator.tbl_name:
-                validators_to_remove.append(validator_id)
-        except Exception:
-            pass
-
-    removed_validators = 0
-    for validator_id in validators_to_remove:
-        del app_ctx.active_validators[validator_id]
-        removed_validators += 1
-
     message = f"DataFrame '{df_id}' deleted successfully."
-    if removed_validators > 0:
-        message += f" Also removed {removed_validators} associated validator(s)."
-
     await ctx.report_progress(100, 100, message)
-
-    return {"status": "success", "message": message, "removed_validators": removed_validators}
-
-
-@mcp.tool(
-    name="delete_validator",
-    description="Remove a validator from the server context.",
-    tags={"Validation"},
-)
-async def delete_validator(
-    ctx: Context,
-    validator_id: Annotated[str, "ID of the validator to delete."],
-) -> Dict[str, str]:
-    """
-    Removes a validator from the server context.
-    """
-    app_ctx: AppContext = ctx.request_context.lifespan_context
-
-    if validator_id not in app_ctx.active_validators:
-        raise ValueError(f"Validator ID '{validator_id}' not found.")
-
-    # Get validator info before deleting
-    validator = app_ctx.active_validators[validator_id]
-    table_name = getattr(validator, "tbl_name", "Unknown")
-
-    # Remove the validator
-    del app_ctx.active_validators[validator_id]
-
-    message = f"Validator '{validator_id}' (table: {table_name}) deleted successfully."
-    await ctx.report_progress(100, 100, message)
-
     return {"status": "success", "message": message}
 
 
-def clean_for_json_serialization(obj: Any) -> Any:
-    """
-    Recursively clean an object to ensure it can be JSON serialized by converting
-    problematic values like NaN and infinity.
-    """
-    if isinstance(obj, float):
-        if math.isnan(obj):
-            return None
-        elif math.isinf(obj):
-            return str(obj)  # "inf" or "-inf"
-        else:
-            return obj
-    elif isinstance(obj, dict):
-        return {key: clean_for_json_serialization(value) for key, value in obj.items()}
-    elif isinstance(obj, (list, tuple)):
-        return [clean_for_json_serialization(item) for item in obj]
-    else:
-        return obj
+# =============================================================================
+# TOOLS: Data Analysis
+# =============================================================================
+
+
+@mcp.tool(
+    name="analyze_data_quality",
+    description="Analyze data quality using Pointblank's DataScan class.",
+    tags={"Data Analysis"},
+)
+async def analyze_data_quality(
+    ctx: Context,
+    df_id: Annotated[str, "ID of the DataFrame to analyze."],
+) -> Dict[str, Any]:
+    """Analyze data quality using Pointblank's built-in DataScan functionality."""
+    app_ctx: AppContext = ctx.request_context.lifespan_context
+    df_id = validate_resource_id(df_id, "DataFrame")
+
+    if df_id not in app_ctx.loaded_dataframes:
+        raise ValueError(f"DataFrame ID '{df_id}' not found.")
+
+    df = app_ctx.loaded_dataframes[df_id]
+
+    await ctx.report_progress(20, 100, "Starting data quality analysis...")
+
+    scanner = pb.DataScan(data=df)
+
+    await ctx.report_progress(60, 100, "Running DataScan analysis...")
+
+    profile_json = scanner.to_json()
+
+    await ctx.report_progress(80, 100, "Processing results...")
+
+    profile_results = json.loads(profile_json)
+    cleaned_results = clean_for_json_serialization(profile_results)
+
+    await ctx.report_progress(100, 100, "Data quality analysis complete!")
+
+    return {"status": "success", "df_id": df_id, "analysis": cleaned_results}
 
 
 @mcp.tool(
@@ -899,10 +346,9 @@ async def profile_dataframe(
         int, "Maximum number of rows to sample for profiling (0 = all rows)."
     ] = 10000,
 ) -> Dict[str, Any]:
-    """
-    Generates comprehensive data profiling insights using Pointblank's DataScan class.
-    """
+    """Generates comprehensive data profiling insights using Pointblank's DataScan class."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
+    df_id = validate_resource_id(df_id, "DataFrame")
 
     if df_id not in app_ctx.loaded_dataframes:
         raise ValueError(f"DataFrame ID '{df_id}' not found.")
@@ -911,11 +357,10 @@ async def profile_dataframe(
 
     await ctx.report_progress(10, 100, "Starting data profiling...")
 
-    # Sample data if needed
     if sample_size > 0 and df.shape[0] > sample_size:
-        if hasattr(df, "sample"):  # pandas
+        if hasattr(df, "sample"):
             df_sample = df.sample(n=sample_size, random_state=23)
-        else:  # polars
+        else:
             df_sample = df.sample(n=sample_size, seed=23)
         await ctx.report_progress(20, 100, f"Sampling {sample_size} rows for analysis...")
     else:
@@ -923,444 +368,78 @@ async def profile_dataframe(
 
     await ctx.report_progress(50, 100, "Running Pointblank DataScan...")
 
-    try:
-        # Use Pointblank's DataScan class for profiling
-        scanner = pb.DataScan(data=df_sample)
+    scanner = pb.DataScan(data=df_sample)
 
-        await ctx.report_progress(80, 100, "Converting to JSON...")
+    await ctx.report_progress(80, 100, "Converting to JSON...")
 
-        # Get the JSON representation which properly handles numpy types
-        profile_json = scanner.to_json()
+    profile_json = scanner.to_json()
+    profile_results = json.loads(profile_json)
+    cleaned_results = clean_for_json_serialization(profile_results)
 
-        # Parse it back to a dictionary and clean any problematic values
-        profile_results = json.loads(profile_json)
-        cleaned_results = clean_for_json_serialization(profile_results)
+    await ctx.report_progress(100, 100, "Data profiling complete!")
+    return cleaned_results
 
-        await ctx.report_progress(100, 100, "Data profiling complete!")
 
-        return cleaned_results
-
-    except Exception as e:
-        error_msg = f"Error during data profiling: {str(e)}"
-        await ctx.report_progress(100, 100, error_msg)
-        raise ValueError(error_msg)
+# =============================================================================
+# TOOLS: Validation
+# =============================================================================
 
 
 @mcp.tool(
-    name="apply_validation_template",
-    description="Apply a pre-built validation template to a validator for common data quality patterns.",
+    name="list_active_validators",
+    description="List all validators currently active in the server context.",
     tags={"Validation"},
 )
-async def apply_validation_template(
-    ctx: Context,
-    validator_id: Annotated[str, "ID of the Validator to apply template to."],
-    template_name: Annotated[
-        str,
-        "Name of the validation template to apply. Available: 'basic_quality', 'financial_data', 'customer_data', 'sensor_data', 'survey_data'.",
-    ],
-    column_mapping: Annotated[
-        Dict[str, str],
-        "Mapping of template column names to actual DataFrame column names. E.g., {'amount': 'price', 'id': 'customer_id'}.",
-    ],
-    thresholds: Annotated[
-        Optional[Dict[str, float]],
-        "Optional custom thresholds for validation steps. E.g., {'warning': 0.05, 'error': 0.1}.",
-    ] = None,
-) -> Dict[str, Any]:
-    """
-    Applies a pre-built validation template with common data quality checks.
-    Templates include ready-made validation rules for typical data scenarios.
-    """
+async def list_active_validators(ctx: Context) -> Dict[str, Any]:
+    """Returns information about all active validators in the server."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
+
+    validators_info = {}
+    for validator_id, validator in app_ctx.active_validators.items():
+        try:
+            table_name = getattr(validator, "tbl_name", "Unknown")
+            label = getattr(validator, "label", "No label")
+            is_interrogated = (
+                hasattr(validator, "time_processed") and validator.time_processed is not None
+            )
+            step_count = len(getattr(validator, "_validation_set", []))
+
+            validators_info[validator_id] = {
+                "table_name": table_name,
+                "label": label,
+                "is_interrogated": is_interrogated,
+                "validation_steps_count": step_count,
+            }
+        except Exception as e:
+            validators_info[validator_id] = {"error": f"Failed to get info: {str(e)}"}
+
+    return {
+        "active_validators": validators_info,
+        "total_count": len(validators_info),
+    }
+
+
+@mcp.tool(
+    name="delete_validator",
+    description="Remove a validator from the server context.",
+    tags={"Validation"},
+)
+async def delete_validator(
+    ctx: Context,
+    validator_id: Annotated[str, "ID of the validator to delete."],
+) -> Dict[str, str]:
+    """Removes a validator from the server context."""
+    app_ctx: AppContext = ctx.request_context.lifespan_context
+    validator_id = validate_resource_id(validator_id, "Validator")
 
     if validator_id not in app_ctx.active_validators:
         raise ValueError(f"Validator ID '{validator_id}' not found.")
 
-    # Get validation template
-    template = _get_validation_template(template_name)
-    if not template:
-        available_templates = [
-            "basic_quality",
-            "financial_data",
-            "customer_data",
-            "sensor_data",
-            "survey_data",
-        ]
-        raise ValueError(f"Unknown template '{template_name}'. Available: {available_templates}")
+    del app_ctx.active_validators[validator_id]
 
-    validator = app_ctx.active_validators[validator_id]
-    applied_validations = []
-
-    await ctx.report_progress(10, 100, f"Applying {template_name} template...")
-
-    # Apply each validation in the template
-    for i, validation_rule in enumerate(template["validations"]):
-        try:
-            # Map template column names to actual column names
-            mapped_params = {}
-            for key, value in validation_rule["params"].items():
-                if key == "columns" and value in column_mapping:
-                    mapped_params[key] = column_mapping[value]
-                elif isinstance(value, str) and value in column_mapping:
-                    mapped_params[key] = column_mapping[value]
-                else:
-                    mapped_params[key] = value
-
-            # Add custom thresholds if provided
-            if thresholds:
-                mapped_params.update(thresholds)
-
-            # Get the validation method
-            validation_type = validation_rule["validation_type"]
-            supported_validations = _get_supported_validations(validator)
-
-            if validation_type in supported_validations:
-                validation_method = supported_validations[validation_type]
-                validation_method(**mapped_params)
-
-                applied_validations.append(
-                    {
-                        "validation_type": validation_type,
-                        "params": mapped_params,
-                        "description": validation_rule.get("description", ""),
-                    }
-                )
-
-                await ctx.report_progress(
-                    10 + (i + 1) * 80 // len(template["validations"]),
-                    100,
-                    f"Applied {validation_type}...",
-                )
-            else:
-                applied_validations.append(
-                    {
-                        "validation_type": validation_type,
-                        "params": mapped_params,
-                        "error": f"Unsupported validation type: {validation_type}",
-                    }
-                )
-
-        except Exception as e:
-            applied_validations.append(
-                {
-                    "validation_type": validation_rule["validation_type"],
-                    "params": validation_rule["params"],
-                    "error": str(e),
-                }
-            )
-
-    await ctx.report_progress(100, 100, f"Template {template_name} applied successfully!")
-
-    return {
-        "template_name": template_name,
-        "template_description": template["description"],
-        "applied_validations": applied_validations,
-        "total_validations": len(applied_validations),
-        "successful_validations": len([v for v in applied_validations if "error" not in v]),
-    }
-
-
-def _get_validation_template(template_name: str) -> Optional[Dict[str, Any]]:
-    """Get a predefined validation template."""
-    templates = {
-        "basic_quality": {
-            "description": "Basic data quality checks for any dataset",
-            "validations": [
-                {
-                    "validation_type": "col_exists",
-                    "params": {"columns": "id"},
-                    "description": "Check that ID column exists",
-                },
-                {
-                    "validation_type": "col_vals_not_null",
-                    "params": {"columns": "id"},
-                    "description": "ID column should not have null values",
-                },
-                {
-                    "validation_type": "rows_distinct",
-                    "params": {},
-                    "description": "Check for duplicate rows",
-                },
-            ],
-        },
-        "financial_data": {
-            "description": "Validation template for financial/transaction data",
-            "validations": [
-                {
-                    "validation_type": "col_vals_not_null",
-                    "params": {"columns": "amount"},
-                    "description": "Amount should not be null",
-                },
-                {
-                    "validation_type": "col_vals_gt",
-                    "params": {"columns": "amount", "value": 0},
-                    "description": "Amount should be positive",
-                },
-                {
-                    "validation_type": "col_vals_not_null",
-                    "params": {"columns": "date"},
-                    "description": "Transaction date should not be null",
-                },
-                {
-                    "validation_type": "col_vals_not_null",
-                    "params": {"columns": "account_id"},
-                    "description": "Account ID should not be null",
-                },
-            ],
-        },
-        "customer_data": {
-            "description": "Validation template for customer/user data",
-            "validations": [
-                {
-                    "validation_type": "col_vals_not_null",
-                    "params": {"columns": "customer_id"},
-                    "description": "Customer ID should not be null",
-                },
-                {
-                    "validation_type": "col_vals_regex",
-                    "params": {"columns": "email", "pattern": r"^[^@]+@[^@]+\.[^@]+$"},
-                    "description": "Email should be in valid format",
-                },
-                {
-                    "validation_type": "col_vals_between",
-                    "params": {"columns": "age", "left": 0, "right": 120},
-                    "description": "Age should be between 0 and 120",
-                },
-                {
-                    "validation_type": "col_vals_in_set",
-                    "params": {"columns": "status", "set_": ["active", "inactive", "pending"]},
-                    "description": "Status should be one of predefined values",
-                },
-            ],
-        },
-        "sensor_data": {
-            "description": "Validation template for IoT/sensor data",
-            "validations": [
-                {
-                    "validation_type": "col_vals_not_null",
-                    "params": {"columns": "timestamp"},
-                    "description": "Timestamp should not be null",
-                },
-                {
-                    "validation_type": "col_vals_not_null",
-                    "params": {"columns": "sensor_id"},
-                    "description": "Sensor ID should not be null",
-                },
-                {
-                    "validation_type": "col_vals_between",
-                    "params": {"columns": "temperature", "left": -50, "right": 100},
-                    "description": "Temperature should be in reasonable range",
-                },
-                {
-                    "validation_type": "col_vals_ge",
-                    "params": {"columns": "humidity", "value": 0},
-                    "description": "Humidity should be non-negative",
-                },
-                {
-                    "validation_type": "col_vals_le",
-                    "params": {"columns": "humidity", "value": 100},
-                    "description": "Humidity should not exceed 100%",
-                },
-            ],
-        },
-        "survey_data": {
-            "description": "Validation template for survey/questionnaire data",
-            "validations": [
-                {
-                    "validation_type": "col_vals_not_null",
-                    "params": {"columns": "response_id"},
-                    "description": "Response ID should not be null",
-                },
-                {
-                    "validation_type": "col_vals_between",
-                    "params": {"columns": "satisfaction_score", "left": 1, "right": 10},
-                    "description": "Satisfaction score should be between 1 and 10",
-                },
-                {
-                    "validation_type": "col_vals_in_set",
-                    "params": {
-                        "columns": "completion_status",
-                        "set_": ["complete", "partial", "abandoned"],
-                    },
-                    "description": "Completion status should be one of predefined values",
-                },
-            ],
-        },
-    }
-
-    return templates.get(template_name)
-
-
-def _get_supported_validations(validator):
-    """Get the supported validation methods from a validator instance."""
-    return {
-        # Column value validations
-        "col_vals_lt": validator.col_vals_lt,
-        "col_vals_gt": validator.col_vals_gt,
-        "col_vals_le": validator.col_vals_le,
-        "col_vals_ge": validator.col_vals_ge,
-        "col_vals_eq": validator.col_vals_eq,
-        "col_vals_ne": validator.col_vals_ne,
-        "col_vals_between": validator.col_vals_between,
-        "col_vals_outside": validator.col_vals_outside,
-        "col_vals_in_set": validator.col_vals_in_set,
-        "col_vals_not_in_set": validator.col_vals_not_in_set,
-        "col_vals_null": validator.col_vals_null,
-        "col_vals_not_null": validator.col_vals_not_null,
-        "col_vals_regex": validator.col_vals_regex,
-        "col_vals_expr": validator.col_vals_expr,
-        "col_count_match": validator.col_count_match,
-        # Check existence of a column
-        "col_exists": validator.col_exists,
-        # Row validations
-        "rows_distinct": validator.rows_distinct,
-        "rows_complete": validator.rows_complete,
-        "row_count_match": validator.row_count_match,
-        # Other specialized validations
-        "conjointly": validator.conjointly,
-        "col_schema_match": validator.col_schema_match,
-    }
-
-
-@mcp.tool(
-    name="server_health_check",
-    description="Get comprehensive server health and status information.",
-    tags={"Server Management"},
-)
-async def server_health_check(ctx: Context) -> Dict[str, Any]:
-    """
-    Returns comprehensive server health information including:
-    - resource usage and capacity
-    - backend availability
-    - active resources count
-    - system information
-    """
-    app_ctx: AppContext = ctx.request_context.lifespan_context
-
-    # Get current time
-    current_time = datetime.now().isoformat()
-
-    # Count resources
-    dataframes_count = len(app_ctx.loaded_dataframes)
-    validators_count = len(app_ctx.active_validators)
-
-    # Calculate memory usage
-    total_memory_mb = 0
-    dataframe_details = []
-
-    for df_id, df in app_ctx.loaded_dataframes.items():
-        try:
-            if hasattr(df, "memory_usage"):  # pandas
-                memory_mb = df.memory_usage(deep=True).sum() / 1024 / 1024
-                total_memory_mb += memory_mb
-                dataframe_details.append(
-                    {
-                        "df_id": df_id,
-                        "shape": [int(df.shape[0]), int(df.shape[1])],  # Convert to Python ints
-                        "memory_mb": round(memory_mb, 2),
-                        "backend": "pandas",
-                    }
-                )
-            else:  # polars or other
-                # Estimate memory for polars (rough approximation)
-                estimated_mb = (
-                    df.shape[0] * df.shape[1] * 8 / 1024 / 1024
-                )  # 8 bytes per value estimate
-                total_memory_mb += estimated_mb
-                dataframe_details.append(
-                    {
-                        "df_id": df_id,
-                        "shape": [int(df.shape[0]), int(df.shape[1])],  # Convert to Python ints
-                        "memory_mb": round(estimated_mb, 2),
-                        "backend": "polars/other",
-                    }
-                )
-        except Exception as e:
-            dataframe_details.append({"df_id": df_id, "error": str(e), "backend": "unknown"})
-
-    # Get system information
-    health_info = {
-        "timestamp": current_time,
-        "server_status": "healthy",
-        "system_info": {
-            "python_version": sys.version.split()[0],
-            "platform": platform.platform(),
-            "architecture": platform.architecture()[0],
-        },
-        "backend_status": {
-            "pandas_available": HAS_PANDAS,
-            "polars_available": HAS_POLARS,
-            "fastmcp_loaded": True,  # If we're running, FastMCP is loaded
-            "pointblank_loaded": True,  # If we're running, Pointblank is loaded
-        },
-        "resource_usage": {
-            "total_dataframes": dataframes_count,
-            "total_validators": validators_count,
-            "total_memory_mb": round(total_memory_mb, 2),
-            "dataframe_details": dataframe_details,
-        },
-        "capabilities": {
-            "supported_file_formats": ["CSV", "JSON", "JSONL", "Parquet"]
-            + (["Excel"] if HAS_PANDAS else []),
-            "validation_types_count": len(_get_supported_validations_list()),
-            "templates_available": [
-                "basic_quality",
-                "financial_data",
-                "customer_data",
-                "sensor_data",
-                "survey_data",
-            ],
-        },
-    }
-
-    # Add warnings if any
-    warnings = []
-    if not HAS_PANDAS and not HAS_POLARS:
-        warnings.append("No DataFrame backends available")
-        health_info["server_status"] = "degraded"
-
-    if total_memory_mb > 1000:  # > 1GB
-        warnings.append(f"High memory usage: {total_memory_mb:.1f}MB")
-
-    if dataframes_count > 50:
-        warnings.append(f"Many DataFrames loaded: {dataframes_count}")
-
-    if warnings:
-        health_info["warnings"] = warnings
-
-    logger.info(
-        f"Health check performed: {dataframes_count} DataFrames, {validators_count} validators, {total_memory_mb:.1f}MB memory"
-    )
-
-    return health_info
-
-
-def _get_supported_validations_list():
-    """Get list of supported validation types."""
-    return [
-        "col_vals_lt",
-        "col_vals_gt",
-        "col_vals_le",
-        "col_vals_ge",
-        "col_vals_eq",
-        "col_vals_ne",
-        "col_vals_between",
-        "col_vals_outside",
-        "col_vals_in_set",
-        "col_vals_not_in_set",
-        "col_vals_null",
-        "col_vals_not_null",
-        "col_vals_regex",
-        "col_vals_expr",
-        "col_count_match",
-        "col_exists",
-        "rows_distinct",
-        "rows_complete",
-        "prompt",
-        "row_count_match",
-        "conjointly",
-        "col_schema_match",
-    ]
+    message = f"Validator '{validator_id}' deleted successfully."
+    await ctx.report_progress(100, 100, message)
+    return {"status": "success", "message": message}
 
 
 @dataclass
@@ -1376,40 +455,20 @@ class ValidatorInfo:
 def create_validator(
     ctx: Context,
     df_id: Annotated[str, "ID of the DataFrame to validate."],
-    validator_id: Annotated[
-        Optional[str],
-        "Optional ID for the Validator. If not provided, a new ID will be generated.",
-    ] = None,
-    table_name: Annotated[
-        Optional[str],
-        "Optional name for the table within Pointblank reports. If not provided, a default name will be used.",
-    ] = None,
-    validator_label: Annotated[
-        Optional[str],
-        "Optional descriptive label for the Validator. If not provided, a default label will be used.",
-    ] = None,  # Corresponds to 'label' in pb.Validate
+    validator_id: Annotated[Optional[str], "Optional ID for the Validator."] = None,
+    table_name: Annotated[Optional[str], "Optional name for the table within reports."] = None,
+    validator_label: Annotated[Optional[str], "Optional descriptive label."] = None,
     thresholds_dict: Annotated[
         Optional[Dict[str, Union[int, float]]],
-        "Optional thresholds for validation failures. Example: {'warning': 0.1, 'error': 5, 'critical': 0.10}. "
-        "If not provided, no thresholds will be set.",
-    ] = None,  # Corresponds to 'thresholds' in pb.Validate, e.g. {"warning": 1, "error": 20, "critical": 0.10}
-    actions_dict: Optional[Dict[str, Any]] = None,  # Simplified, for pb.Actions
-    final_actions_dict: Optional[Dict[str, Any]] = None,  # Simplified, for pb.FinalActions
+        "Optional thresholds. Example: {'warning': 0.1, 'error': 5, 'critical': 0.10}.",
+    ] = None,
     brief: Optional[bool] = None,
     lang: Optional[str] = None,
     locale: Optional[str] = None,
 ) -> ValidatorInfo:
-    """
-    Creates a Pointblank Validator for a previously loaded DataFrame.
-    Assigns a unique ID to the Validator for adding validation steps.
-    If validator_id is not provided, a new one will be generated.
-    'df_id' must refer to a DataFrame loaded via 'load_dataframe'.
-    'table_name' is an optional name for the table within Pointblank reports.
-    'validator_label' is an optional descriptive label for the validator.
-    'thresholds_dict' can be like {"warning": 0.1, "error": 5} to set failure thresholds.
-    Returns the Validator ID.
-    """
+    """Creates a Pointblank Validator for a previously loaded DataFrame."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
+    df_id = validate_resource_id(df_id, "DataFrame")
 
     if df_id not in app_ctx.loaded_dataframes:
         raise ValueError(
@@ -1419,66 +478,30 @@ def create_validator(
     df = app_ctx.loaded_dataframes[df_id]
 
     effective_validator_id = validator_id if validator_id else f"validator_{uuid.uuid4().hex[:8]}"
+    effective_validator_id = validate_resource_id(effective_validator_id, "Validator")
 
     if effective_validator_id in app_ctx.active_validators:
-        raise ValueError(
-            f"Validator ID '{effective_validator_id}' already exists. Choose a different ID or omit to generate a new one."
-        )
+        raise ValueError(f"Validator ID '{effective_validator_id}' already exists.")
 
     actual_table_name = table_name if table_name else f"table_for_{df_id}"
-    actual_validator_label = (
-        validator_label if validator_label else f"Validation for {actual_table_name}"
-    )
+    actual_label = validator_label if validator_label else f"Validation for {actual_table_name}"
 
-    # Construct Thresholds, Actions, FinalActions if dicts are provided
-    pb_thresholds = None
-    if thresholds_dict:
-        try:
-            pb_thresholds = pb.Thresholds(**thresholds_dict)
-        except Exception as e:
-            raise ValueError(f"Error creating pb.Thresholds from thresholds_dict: {e}")
-
-    # Note: pb.Actions and pb.FinalActions might require more complex construction
-    # For simplicity, we're assuming direct kwarg passing or simple structures.
-    # This part might need refinement based on how pb.Actions/pb.FinalActions are instantiated.
-    pb_actions = None
-    if actions_dict:
-        try:
-            # Example: if pb.Actions takes specific function handlers
-            # This is a placeholder and likely needs more specific handling
-            pb_actions = pb.Actions(
-                **actions_dict
-            )  # This assumes pb.Actions can be created this way
-        except Exception as e:
-            print(f"Could not create pb.Actions from actions_dict: {e}. Passing None.")
-
-    pb_final_actions = None
-    if final_actions_dict:
-        try:
-            pb_final_actions = pb.FinalActions(**final_actions_dict)  # Placeholder
-        except Exception as e:
-            print(f"Could not create pb.FinalActions from final_actions_dict: {e}. Passing None.")
-
-    validator_instance_params = {
+    validator_params = {
         "data": df,
         "tbl_name": actual_table_name,
-        "label": actual_validator_label,
+        "label": actual_label,
     }
 
-    if pb_thresholds:
-        validator_instance_params["thresholds"] = pb_thresholds
-    if pb_actions:
-        validator_instance_params["actions"] = pb_actions
-    if pb_final_actions:
-        validator_instance_params["final_actions"] = pb_final_actions
+    if thresholds_dict:
+        validator_params["thresholds"] = pb.Thresholds(**thresholds_dict)
     if brief is not None:
-        validator_instance_params["brief"] = brief
+        validator_params["brief"] = brief
     if lang:
-        validator_instance_params["lang"] = lang
+        validator_params["lang"] = lang
     if locale:
-        validator_instance_params["locale"] = locale
+        validator_params["locale"] = locale
 
-    validator_instance = pb.Validate(**validator_instance_params)
+    validator_instance = pb.Validate(**validator_params)
     app_ctx.active_validators[effective_validator_id] = validator_instance
 
     return ValidatorInfo(validator_id=effective_validator_id)
@@ -1500,107 +523,120 @@ def add_validation_step(
     validator_id: Annotated[str, "ID of the Validator to add a step to."],
     validation_type: Annotated[
         str,
-        "Type of validation to perform. Supported types include: 'col_vals_lt', 'col_vals_gt', 'col_vals_between', 'col_exists', 'rows_distinct', etc.",
+        "Type of validation to perform (e.g., 'col_vals_lt', 'col_vals_between', 'rows_distinct').",
     ],
     params: Annotated[
         Dict[str, Any],
-        "Parameters for the validation function. This should match the expected parameters for the Pointblank validation method.",
+        "Parameters for the validation function matching Pointblank's API.",
     ],
-    actions_config: Optional[Dict[str, Any]] = None,  # Placeholder for simplified action definition
 ) -> ValidationStepInfo:
-    """
-    Adds a validation step to an existing Pointblank Validator.
-    'validator_id' must refer to a validator created via 'create_validator'.
-    'validation_type' specifies the Pointblank validation function to call
-      (e.g., 'col_vals_lt', 'col_vals_between', 'col_vals_in_set', 'col_exists', 'rows_distinct').
-    'params' is a dictionary of parameters for that validation function.
-    'actions_config' (optional) can be used to define simple actions (currently basic support).
-    """
+    """Adds a validation step to an existing Pointblank Validator."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
+    validator_id = validate_resource_id(validator_id, "Validator")
 
     if validator_id not in app_ctx.active_validators:
-        raise ValueError(
-            f"Validator ID '{validator_id}' not found. Please create it first using 'create_validator'."
-        )
+        raise ValueError(f"Validator ID '{validator_id}' not found.")
 
     validator = app_ctx.active_validators[validator_id]
+    supported = get_supported_validations(validator)
 
-    # --- Define supported validation types and their methods from pb.Validate ---
-    # This mapping allows dynamic dispatch and can be extended
-    # Methods are called on the 'validator' (pb.Validate instance)
-    supported_validations = {
-        # Column value validations
-        "col_vals_lt": validator.col_vals_lt,  # less than a value
-        "col_vals_gt": validator.col_vals_gt,  # greater than a value
-        "col_vals_le": validator.col_vals_le,  # less or equal
-        "col_vals_ge": validator.col_vals_ge,  # greater or equal
-        "col_vals_eq": validator.col_vals_eq,  # equal to a value
-        "col_vals_ne": validator.col_vals_ne,  # not equal to a value
-        "col_vals_between": validator.col_vals_between,  # data lies between two values left=val, right=val
-        "col_vals_outside": validator.col_vals_outside,  # data is outside two values
-        "col_vals_in_set": validator.col_vals_in_set,  # values in a set e.g. [1,2,3]
-        "col_vals_not_in_set": validator.col_vals_not_in_set,  # values not in a set
-        "col_vals_null": validator.col_vals_null,  # null values
-        "col_vals_not_null": validator.col_vals_not_null,  # not null values
-        "col_vals_regex": validator.col_vals_regex,  # values match a regular expresion
-        "col_vals_expr": validator.col_vals_expr,  # Validate column values using a custom expression
-        "col_count_match": validator.col_count_match,  # Validate whether the column count of the table matches a specified count.
-        # Check existence of a column
-        "col_exists": validator.col_exists,
-        # Row validations
-        "rows_distinct": validator.rows_distinct,  # check distinct rows in a table
-        "rows_complete": validator.rows_complete,  # check for no nulls in rows across specified columns
-        "prompt": validator.prompt,  # AI-powered validation of rows using LLMs
-        "row_count_match": validator.row_count_match,  # check if number of rows in the table matches a fixed value
-        # Other specialized validations
-        "conjointly": validator.conjointly,  # For multiple column conditions
-        "col_schema_match": validator.col_schema_match,  # Do columns in the table (and their types) match a predefined schema?
-    }
-
-    if validation_type not in supported_validations:
+    if validation_type not in supported:
         raise ValueError(
-            f"Unsupported validation_type: '{validation_type}'. Supported types include: {list(supported_validations.keys())}"
+            f"Unsupported validation_type: '{validation_type}'. Supported: {list(supported.keys())}"
         )
 
-    validation_method = supported_validations[validation_type]
-
-    # Simplified actions handling (can be expanded)
-    # pb.Validate methods expect an 'actions' parameter which is an instance of pb.Actions
-    # This is a placeholder for how one might construct it.
-    # A more robust solution would deserialize a dict into pb.Actions object.
     current_params = {**params}
-
-    # Handle parameter name mapping for reserved keywords
-    # Pointblank uses 'set' but we use 'set_' in JSON to avoid Python reserved keyword issues
+    # Handle 'set_' -> 'set' mapping for Python reserved keyword
     if "set_" in current_params:
         current_params["set"] = current_params.pop("set_")
 
-    if actions_config:
-        # Example: actions_config = {"warn": 0.1} might translate to
-        # actions = pb.Actions(warn_fraction=0.1)
-        # For now, if a method expects 'actions', it should be in params directly
-        # or handled here explicitly if simple shorthands are desired.
-        # This is a complex area to generalize perfectly via JSON.
-        # Let's assume 'actions' if needed is part of 'params' and is a pb.Actions object
-        # or the LLM constructs the params for methods that take thresholds directly.
-        # For now, we'll pass 'params' as is.
-        # If 'actions' is a direct parameter of the validation_method, it should be in 'params'.
-        pass  # No special action processing here yet, assuming 'params' has all needed args
     try:
-        validation_method(**current_params)
+        supported[validation_type](**current_params)
     except TypeError as e:
-        raise ValueError(
-            f"Error calling validation method '{validation_type}' with params {current_params}. Original error: {e}. Check parameter names and types against Pointblank's API for the '{validation_type}' method of the 'Validate' class."
-        )
-    except Exception as e:
-        raise RuntimeError(
-            f"An unexpected error occurred while adding validation step '{validation_type}': {e}"
-        )
+        raise ValueError(f"Error calling '{validation_type}' with params {current_params}: {e}")
 
     return ValidationStepInfo(
         validator_id=validator_id,
         status=f"Validation step '{validation_type}' added successfully.",
     )
+
+
+@mcp.tool(
+    name="apply_validation_template",
+    description="Apply a pre-built validation template to a validator.",
+    tags={"Validation"},
+)
+async def apply_validation_template(
+    ctx: Context,
+    validator_id: Annotated[str, "ID of the Validator to apply template to."],
+    template_name: Annotated[
+        str,
+        f"Template name. Available: {', '.join(AVAILABLE_TEMPLATES)}.",
+    ],
+    column_mapping: Annotated[
+        Dict[str, str],
+        "Mapping of template column names to actual DataFrame column names.",
+    ],
+) -> Dict[str, Any]:
+    """Applies a pre-built validation template with common data quality checks."""
+    app_ctx: AppContext = ctx.request_context.lifespan_context
+    validator_id = validate_resource_id(validator_id, "Validator")
+
+    if validator_id not in app_ctx.active_validators:
+        raise ValueError(f"Validator ID '{validator_id}' not found.")
+
+    template = get_validation_template(template_name)
+    if not template:
+        raise ValueError(f"Unknown template '{template_name}'. Available: {AVAILABLE_TEMPLATES}")
+
+    validator = app_ctx.active_validators[validator_id]
+    supported = get_supported_validations(validator)
+    applied_validations = []
+
+    await ctx.report_progress(10, 100, f"Applying {template_name} template...")
+
+    for i, rule in enumerate(template["validations"]):
+        try:
+            mapped_params = {}
+            for key, value in rule["params"].items():
+                if key == "columns" and value in column_mapping:
+                    mapped_params[key] = column_mapping[value]
+                elif isinstance(value, str) and value in column_mapping:
+                    mapped_params[key] = column_mapping[value]
+                else:
+                    mapped_params[key] = value
+
+            vtype = rule["validation_type"]
+            if vtype in supported:
+                supported[vtype](**mapped_params)
+                applied_validations.append(
+                    {
+                        "validation_type": vtype,
+                        "params": mapped_params,
+                        "description": rule.get("description", ""),
+                    }
+                )
+
+            await ctx.report_progress(
+                10 + (i + 1) * 80 // len(template["validations"]), 100, f"Applied {vtype}..."
+            )
+        except Exception as e:
+            applied_validations.append(
+                {
+                    "validation_type": rule["validation_type"],
+                    "error": str(e),
+                }
+            )
+
+    await ctx.report_progress(100, 100, f"Template {template_name} applied!")
+
+    return {
+        "template_name": template_name,
+        "template_description": template["description"],
+        "applied_validations": applied_validations,
+        "total_validations": len(applied_validations),
+        "successful_validations": len([v for v in applied_validations if "error" not in v]),
+    }
 
 
 @dataclass
@@ -1618,83 +654,58 @@ class ValidationOutput:
 async def get_validation_step_output(
     ctx: Context,
     validator_id: Annotated[str, "ID of the Validator to retrieve output from."],
-    output_path: Annotated[
-        str,
-        "Path to save the output file. Must end with .csv.",
-    ],
-    sundered_type: Annotated[
-        str,
-        "Mode 2: Retrieve all 'pass' or 'fail' rows for the *entire* validation run. Only used if 'step_index' is not provided.",
-    ] = "fail",
+    output_path: Annotated[str, "Path to save the output file. Must end with .csv."],
+    sundered_type: Annotated[str, "Retrieve all 'pass' or 'fail' rows."] = "fail",
     step_index: Annotated[
-        Optional[int],
-        "Mode 1: Retrieve data for a *specific* step by its index (starting from 0). If used, 'sundered_type' is ignored.",
+        Optional[int], "Specific step index (0-based). Overrides sundered_type."
     ] = None,
 ) -> ValidationOutput:
-    """
-    Retrieves validation output and saves it to a CSV file. This function has two modes:
-    1.  Specific Step Extract: Provide a 'step_index' to get the data extract (e.g., failing rows) for that specific step.
-    2.  Overall Sundered Data: Omit 'step_index' and use 'sundered_type' ('pass' or 'fail') to get all rows that met that condition across all validation steps.
-    """
+    """Retrieves validation output and saves it to a CSV file."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
+    validator_id = validate_resource_id(validator_id, "Validator")
 
     if validator_id not in app_ctx.active_validators:
         raise ValueError(f"Validator ID '{validator_id}' not found.")
     validator = app_ctx.active_validators[validator_id]
 
-    p_output_path = _validate_output_path(output_path, {".csv"})
+    p_output_path = validate_output_path(output_path, {".csv"})
 
     if step_index is not None and step_index < 0:
         raise ValueError("The 'step_index' cannot be a negative number.")
 
-    try:
-        if not getattr(validator, "time_processed", None):
-            await ctx.warning(
-                f"Validator '{validator_id}' has not been interrogated. Interrogating now."
-            )
-
-        message = ""
-        data_extract_df = None
-
-        # Pathway 1: Get data for a single, specific validation step.
-        if step_index is not None:
-            data_extract_df = validator.get_data_extracts(i=step_index, frame=True)
-            if data_extract_df is None or data_extract_df.empty:
-                message = f"No data extract available for step {step_index}. This may mean all rows passed this validation step."
-                data_extract_df = None  # Ensure it's None if empty
-            else:
-                message = f"Data extract for step {step_index} retrieved."
-
-        # Pathway 2: Get all 'fail' or 'pass' data from the entire validation run.
-        else:
-            data_extract_df = validator.get_sundered_data(type=sundered_type)
-            if data_extract_df is None or data_extract_df.empty:
-                message = f"No sundered data available for type '{sundered_type}'."
-                data_extract_df = None  # Ensure it's None if empty
-            else:
-                message = f"Sundered data for type '{sundered_type}' retrieved."
-
-        if data_extract_df is None:
-            return ValidationOutput(
-                status="success",
-                message=message,
-                output_file=None,
-            )
-
-        # Save to CSV using backend-agnostic method
-        _save_dataframe_to_csv(data_extract_df, p_output_path)
-        message = f"Data extract saved to {p_output_path.resolve()}"
-
-        await ctx.report_progress(100, 100, message)
-
-        return ValidationOutput(
-            status="success",
-            message=message,
-            output_file=str(p_output_path.resolve()),
+    if not getattr(validator, "time_processed", None):
+        await ctx.warning(
+            f"Validator '{validator_id}' has not been interrogated. Interrogating now."
         )
 
-    except Exception as e:
-        raise RuntimeError(f"Error getting output for validator '{validator_id}': {e}")
+    message = ""
+    data_extract_df = None
+
+    if step_index is not None:
+        data_extract_df = validator.get_data_extracts(i=step_index, frame=True)
+        if data_extract_df is None or data_extract_df.empty:
+            message = f"No data extract available for step {step_index}."
+            data_extract_df = None
+        else:
+            message = f"Data extract for step {step_index} retrieved."
+    else:
+        data_extract_df = validator.get_sundered_data(type=sundered_type)
+        if data_extract_df is None or data_extract_df.empty:
+            message = f"No sundered data available for type '{sundered_type}'."
+            data_extract_df = None
+        else:
+            message = f"Sundered data for type '{sundered_type}' retrieved."
+
+    if data_extract_df is None:
+        return ValidationOutput(status="success", message=message, output_file=None)
+
+    save_dataframe_to_csv(data_extract_df, p_output_path)
+    message = f"Data extract saved to {p_output_path.resolve()}"
+    await ctx.report_progress(100, 100, message)
+
+    return ValidationOutput(
+        status="success", message=message, output_file=str(p_output_path.resolve())
+    )
 
 
 @mcp.tool(
@@ -1706,233 +717,51 @@ async def interrogate_validator(
     ctx: Context,
     validator_id: Annotated[str, "ID of the Validator to interrogate."],
 ) -> Dict[str, Any]:
-    """
-    Runs validations and returns a JSON summary.
-    Also generates an HTML report table that opens in browser.
-    Provides Python code equivalent for reproducing the validation.
-    """
+    """Runs validations and returns a JSON summary with HTML report and Python code."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
+    validator_id = validate_resource_id(validator_id, "Validator")
 
     if validator_id not in app_ctx.active_validators:
         raise ValueError(f"Validator ID '{validator_id}' not found.")
 
     validator = app_ctx.active_validators[validator_id]
 
+    validator.interrogate()
+    json_report_str = validator.get_json_report()
+
+    html_report_path = None
     try:
-        validator.interrogate()
-        json_report_str = validator.get_json_report()
+        html_report_path = generate_validation_report_html(validator, validator_id)
+        open_browser_conditionally(f"file://{html_report_path}")
+        await ctx.report_progress(50, 100, "Validation report opened in browser")
+    except Exception as html_error:
+        logger.warning(f"Could not generate HTML report: {html_error}")
 
-        # Generate HTML report table and open in browser
-        try:
-            html_report_path = _generate_validation_report_html(validator, validator_id)
-            _open_browser_conditionally(f"file://{html_report_path}")
-            await ctx.report_progress(
-                50, 100, f"Validation report opened in browser: {html_report_path}"
-            )
-        except Exception as html_error:
-            logger.warning(f"Could not generate HTML report: {html_error}")
-
-        # Generate Python code equivalent
-        try:
-            python_code = _generate_python_code_for_validator(validator, validator_id)
-            await ctx.report_progress(75, 100, "Generated Python code equivalent for validation")
-        except Exception as code_error:
-            logger.warning(f"Could not generate Python code: {code_error}")
-            python_code = "# Error generating Python code equivalent"
-
-    except Exception as e:
-        raise RuntimeError(f"Error during validator interrogation: {e}")
+    python_code = "# Error generating Python code equivalent"
+    try:
+        python_code = generate_python_code_for_validator(validator, validator_id)
+        await ctx.report_progress(75, 100, "Generated Python code equivalent")
+    except Exception as code_error:
+        logger.warning(f"Could not generate Python code: {code_error}")
 
     report_data = json.loads(json_report_str)
 
-    # Enhanced output with Python code
-    output_dict = {
+    await ctx.report_progress(100, 100, "Validation complete!")
+
+    return {
         "validation_summary": report_data,
         "python_code": python_code,
-        "html_report_path": html_report_path if "html_report_path" in locals() else None,
+        "html_report_path": html_report_path,
         "instructions": {
             "html_report": "Interactive validation report opened in your browser",
             "python_code": "Use the provided Python code to reproduce this validation in your own scripts",
-            "customization": "Modify the Python code to adjust validation rules, thresholds, or add new checks",
         },
     }
 
-    await ctx.report_progress(100, 100, "Validation complete! Check browser for detailed report.")
 
-    return output_dict
-
-
-@mcp.prompt(
-    name="prompt_load_dataframe",
-    description="Prompt to load a DataFrame from a file into the server's context for validation.",
-    tags={"Data Management"},
-)
-def prompt_load_dataframe(
-    input_path: str = "Path to the input CSV, Excel or Parquet file.",
-    df_id: Optional[
-        str
-    ] = "Optional ID for the DataFrame. If not provided, a new ID will be generated.",
-) -> tuple:
-    return (
-        Message(
-            "I can load your data from a file into my context for validation.",
-            role="assistant",
-        ),
-        Message(
-            f"Please call `load_dataframe` with input_path='{input_path}'. "
-            f"You can optionally provide a `df_id` (e.g., '{df_id}') to name this dataset, "
-            "or I will generate one for you. Make a note of the returned `df_id` for subsequent steps.",
-            role="user",
-        ),
-    )
-
-
-@mcp.prompt(
-    name="prompt_create_validator",
-    description="Prompt to create a Pointblank Validator for a loaded DataFrame.",
-    tags={"Validation"},
-)
-def prompt_create_validator(
-    df_id: Annotated[str, "ID of the DataFrame to validate."] = "df_default",
-    validator_id: Annotated[
-        Optional[str],
-        "Optional ID for the Validator. If not provided, a new ID will be generated.",
-    ] = "validator_default",
-    table_name: Annotated[
-        Optional[str],
-        "Optional name for the table within Pointblank reports. If not provided, a default name will be used.",
-    ] = "data_table",
-    validator_label: Annotated[
-        Optional[str],
-        "Optional descriptive label for the Validator. If not provided, a default label will be used.",
-    ] = "Validator",
-    thresholds_dict_example: Annotated[
-        Optional[Dict[str, Union[int, float]]],
-        "Example thresholds for validation failures. If not provided, a default example will be used.",
-    ] = None,
-) -> tuple:
-    """
-    Prompt guiding the LLM to create a Pointblank Validator object.
-    Includes an example for thresholds_dict.
-    """
-    thresholds_msg_example = (
-        thresholds_dict_example if thresholds_dict_example else {"warning": 0.05, "error": 10}
-    )
-
-    return (
-        Message(
-            "Once your data is loaded (using its `df_id`), I can create a 'Validator' object to define data quality checks.",
-            role="assistant",
-        ),
-        Message(
-            f"Please call `create_validator` using the `df_id` of your loaded data (e.g., '{df_id}').\n"
-            f"You can optionally provide:\n"
-            f"- `validator_id` (e.g., '{validator_id}') to name this validator instance.\n"
-            f"- `table_name` (e.g., '{table_name}') as a reference name for the data table in reports.\n"
-            f"- `validator_label` (e.g., '{validator_label}') for a descriptive label.\n"
-            f"- `thresholds_dict` (e.g., {thresholds_msg_example}) to set global failure thresholds for validation steps.\n"
-            f"- Other optional parameters like `actions_dict`, `final_actions_dict`, `brief`, `lang`, `locale` can also be specified if needed.\n"
-            "Make a note of the returned `validator_id` to use when adding validation steps.",
-            role="user",
-        ),
-    )
-
-
-@mcp.prompt(
-    name="prompt_add_validation_step_example",
-    description="Prompt to add a validation step to a Pointblank Validator.",
-    tags={"Validation"},
-)
-def prompt_add_validation_step_example() -> tuple:
-    return (
-        Message(
-            "I can add various validation steps to your validator. "
-            "You'll need to specify the 'validator_id', 'validation_type', and 'params' for the step. "
-            "For example, to check if values in column 'age' are less than 100 for validator 'validator_123':",
-            role="assistant",
-        ),
-        Message(
-            "Please call `add_validation_step` with validator_id='validator_123', "
-            "validation_type='col_vals_lt', and params={'columns': 'age', 'value': 100}. "
-            "Note: Parameter names within 'params' (like 'columns', 'value', 'left', 'right', 'set_', etc.) must exactly match what the specific Pointblank validation function expects.\n"
-            "Other examples:\n"
-            "- For 'col_vals_between': params={'columns': 'score', 'left': 0, 'right': 100, 'inclusive': [True, True]}\n"
-            "- For 'col_vals_in_set': params={'columns': 'grade', 'set_': ['A', 'B', 'C']} (Note: Pointblank uses 'set_' for this method's list of values)\n"
-            "- For 'col_exists': params={'columns': 'user_id'}\n"
-            "Refer to the Pointblank Python API for the 'Validate' class for available `validation_type` (method names) and their specific `params`.",
-            role="user",
-        ),
-    )
-
-
-@mcp.prompt(
-    name="prompt_get_validation_step_output",
-    description="Prompt to get validation output by specifying either a step index or a sundered type.",
-    tags={"Validation"},
-)
-def prompt_get_validation_step_output(
-    validator_id: Annotated[str, "Example ID of the Validator."] = "validator_123",
-    step_index: Annotated[
-        Optional[int],
-        "Example step index for the first mode of operation.",
-    ] = 0,
-    sundered_type: Annotated[
-        Optional[str], "Example sundered type ('pass' or 'fail') for the second mode of operation."
-    ] = "fail",
-) -> tuple:
-    """
-    Guides the LLM to get a validation output CSV by choosing one of two modes:
-    1.  By a specific step index.
-    2.  By the overall sundered data type ('pass' or 'fail').
-    """
-    return (
-        Message(
-            "I can extract validation data in two different ways. You must choose one: "
-            "either get data for a *specific step* by its index, or get *all passed or failed rows* from the entire validation run.",
-            role="assistant",
-        ),
-        Message(
-            f"Please call the `get_validation_step_output` tool using only **one** of the following mutually exclusive options:\n\n"
-            f"**OPTION 1: Get data for a specific step**\n"
-            f"To get the data extract for step number {step_index}, use the `step_index` parameter. For example:\n"
-            f"`get_validation_step_output(validator_id='{validator_id}', step_index={step_index}, output_path='step_{step_index}_data.csv')`\n\n"
-            f"**OPTION 2: Get all passed or failed data**\n"
-            f"To get all rows that '{sundered_type}' across all validation steps, use the `sundered_type` parameter. For example:\n"
-            f"`get_validation_step_output(validator_id='{validator_id}', sundered_type='{sundered_type}', output_path='all_{sundered_type}_rows.csv')`",
-            role="user",
-        ),
-    )
-
-
-@mcp.prompt(
-    name="prompt_interrogate_validator",
-    description="Prompt to run validations and generate reports with Python code.",
-    tags={"Validation"},
-)
-def prompt_interrogate_validator(
-    validator_id: Annotated[str, "ID of the Validator to interrogate."],
-) -> tuple:
-    """
-    Prompt guiding the LLM to run validations and generate HTML reports with Python code equivalent.
-    """
-    return (
-        Message(
-            "After all desired validation steps have been added to a validator, I can run the interrogation process. This will execute all checks and generate comprehensive reports.",
-            role="assistant",
-        ),
-        Message(
-            f"Please call `interrogate_validator` with the `validator_id` (e.g., '{validator_id}').\n"
-            f"This will:\n"
-            f"• Execute all validation checks and return a JSON summary\n"
-            f"• Generate an interactive HTML report that opens in your browser\n"
-            f"• Provide Python code equivalent for reproducing the validation\n"
-            f"• Give you the flexibility to customize and extend the validation in your own scripts",
-            role="user",
-        ),
-    )
-
-
-# --- Core Pointblank Table Functions ---
+# =============================================================================
+# TOOLS: Table Visualization
+# =============================================================================
 
 
 @mcp.tool()
@@ -1944,17 +773,13 @@ async def preview_table(
     limit: int = 1000,
     show_row_numbers: bool = True,
 ) -> str:
-    """
-    Display a preview of the DataFrame showing rows from top and bottom.
-
-    Uses Pointblank's built-in preview() function to generate a nicely formatted
-    table view with column types and a sample of the data.
-    """
+    """Display a preview of the DataFrame showing rows from top and bottom."""
     try:
         app_ctx: AppContext = ctx.request_context.lifespan_context
+        dataframe_id = validate_resource_id(dataframe_id, "DataFrame")
 
         if dataframe_id not in app_ctx.loaded_dataframes:
-            return f"❌ Error: DataFrame '{dataframe_id}' not found. Load a DataFrame first."
+            return f"Error: DataFrame '{dataframe_id}' not found. Load a DataFrame first."
 
         data = app_ctx.loaded_dataframes[dataframe_id]
 
@@ -1963,86 +788,76 @@ async def preview_table(
         )
 
         html_output = gt_table.as_raw_html()
-        browser_msg = _save_html_and_open(
+        browser_msg = save_html_and_open(
             html_output,
             title=f"DataFrame Preview: {dataframe_id}",
             filename_prefix=f"pointblank_preview_{dataframe_id}",
         )
 
-        return f"✅ Table preview generated successfully!\n\n{browser_msg}\n\nShowing {n_head} head + {n_tail} tail rows from {data.shape[0]:,} total rows with {data.shape[1]} columns."
+        return f"Table preview generated successfully!\n\n{browser_msg}\n\nShowing {n_head} head + {n_tail} tail rows from {data.shape[0]:,} total rows with {data.shape[1]} columns."
 
     except Exception as e:
         logger.error(f"Error creating table preview: {e}")
-        return f"❌ Error creating preview: {str(e)}"
+        return f"Error creating preview: {str(e)}"
 
 
 @mcp.tool()
-async def missing_values_table(
-    ctx: Context,
-    dataframe_id: str,
-) -> str:
-    """
-    Generate a table showing missing values analysis for the DataFrame.
-
-    Uses Pointblank's built-in missing_vals_tbl() function to show
-    missing value patterns and statistics.
-    """
+async def missing_values_table(ctx: Context, dataframe_id: str) -> str:
+    """Generate a table showing missing values analysis for the DataFrame."""
     try:
         app_ctx: AppContext = ctx.request_context.lifespan_context
+        dataframe_id = validate_resource_id(dataframe_id, "DataFrame")
 
         if dataframe_id not in app_ctx.loaded_dataframes:
-            return f"❌ Error: DataFrame '{dataframe_id}' not found. Load a DataFrame first."
+            return f"Error: DataFrame '{dataframe_id}' not found. Load a DataFrame first."
 
         data = app_ctx.loaded_dataframes[dataframe_id]
 
         gt_table = pb.missing_vals_tbl(data)
         html_output = gt_table.as_raw_html()
-        browser_msg = _save_html_and_open(
+        browser_msg = save_html_and_open(
             html_output,
             title=f"Missing Values Analysis: {dataframe_id}",
             filename_prefix=f"pointblank_missing_values_{dataframe_id}",
         )
 
-        return f"✅ Missing values analysis generated successfully!\n\n{browser_msg}\n\nDataset: {data.shape[0]:,} rows × {data.shape[1]} columns"
+        return f"Missing values analysis generated!\n\n{browser_msg}\n\nDataset: {data.shape[0]:,} rows x {data.shape[1]} columns"
 
     except Exception as e:
         logger.error(f"Error creating missing values table: {e}")
-        return f"❌ Error creating missing values analysis: {str(e)}"
+        return f"Error creating missing values analysis: {str(e)}"
 
 
 @mcp.tool()
-async def column_summary_table(
-    ctx: Context,
-    dataframe_id: str,
-    table_name: str = None,
-) -> str:
-    """
-    Generate a comprehensive column-level summary of the DataFrame.
-
-    Uses Pointblank's built-in col_summary_tbl() function to provide detailed
-    statistics including data types, missing values, and descriptive statistics.
-    """
+async def column_summary_table(ctx: Context, dataframe_id: str, table_name: str = None) -> str:
+    """Generate a comprehensive column-level summary of the DataFrame."""
     try:
         app_ctx: AppContext = ctx.request_context.lifespan_context
+        dataframe_id = validate_resource_id(dataframe_id, "DataFrame")
 
         if dataframe_id not in app_ctx.loaded_dataframes:
-            return f"❌ Error: DataFrame '{dataframe_id}' not found. Load a DataFrame first."
+            return f"Error: DataFrame '{dataframe_id}' not found. Load a DataFrame first."
 
         data = app_ctx.loaded_dataframes[dataframe_id]
 
         gt_table = pb.col_summary_tbl(data, tbl_name=table_name if table_name else dataframe_id)
         html_output = gt_table.as_raw_html()
-        browser_msg = _save_html_and_open(
+        browser_msg = save_html_and_open(
             html_output,
             title=f"Column Summary: {dataframe_id}",
             filename_prefix=f"pointblank_column_summary_{dataframe_id}",
         )
 
-        return f"✅ Column summary table generated successfully!\n\n{browser_msg}\n\nDataset: {data.shape[0]:,} rows × {data.shape[1]} columns"
+        return f"Column summary table generated!\n\n{browser_msg}\n\nDataset: {data.shape[0]:,} rows x {data.shape[1]} columns"
 
     except Exception as e:
         logger.error(f"Error creating column summary table: {e}")
-        return f"❌ Error creating column summary: {str(e)}"
+        return f"Error creating column summary: {str(e)}"
+
+
+# =============================================================================
+# TOOLS: AI & Assistant
+# =============================================================================
 
 
 @mcp.tool(
@@ -2053,269 +868,269 @@ async def column_summary_table(
 async def draft_validation_plan(
     ctx: Context,
     dataframe_id: Annotated[str, "ID of the DataFrame to generate validation plan for."],
-    model: Annotated[
-        str,
-        "AI model to use in format 'provider:model' (e.g., 'anthropic:claude-sonnet-4-5', 'openai:gpt-4'). Supported providers: anthropic, openai, ollama, bedrock.",
-    ] = "anthropic:claude-sonnet-4-5",
-    api_key: Annotated[
-        Optional[str],
-        "API key for the model provider. If not provided, will try to load from environment variables or .env file.",
-    ] = None,
+    model: Annotated[str, "AI model in format 'provider:model'."] = "anthropic:claude-sonnet-4-5",
+    api_key: Annotated[Optional[str], "API key for the model provider."] = None,
 ) -> Dict[str, Any]:
-    """
-    Uses Pointblank's DraftValidation class to generate an AI-powered validation plan.
-    This provides a much more sophisticated and data-aware validation strategy than templates.
-
-    The AI analyzes your data structure, types, ranges, and patterns to generate
-    appropriate validation rules automatically.
-    """
+    """Uses Pointblank's DraftValidation to generate an AI-powered validation plan."""
     app_ctx: AppContext = ctx.request_context.lifespan_context
+    dataframe_id = validate_resource_id(dataframe_id, "DataFrame")
 
     if dataframe_id not in app_ctx.loaded_dataframes:
         raise ValueError(f"DataFrame ID '{dataframe_id}' not found.")
 
     df = app_ctx.loaded_dataframes[dataframe_id]
 
-    await ctx.report_progress(
-        10, 100, f"Initializing AI validation plan generation with {model}..."
-    )
+    await ctx.report_progress(10, 100, f"Initializing AI validation plan with {model}...")
 
     try:
-        # Check if DraftValidation dependencies are available
-        try:
-            from pointblank.draft import DraftValidation
-        except ImportError as e:
-            return {
-                "error": "DraftValidation not available",
-                "message": "The DraftValidation feature requires additional dependencies. Install with: pip install pointblank[generate]",
-                "details": str(e),
-            }
-
-        await ctx.report_progress(
-            30, 100, "Analyzing data structure and generating validation plan..."
-        )
-
-        # Generate the validation plan using AI
-        draft_validator = DraftValidation(data=df, model=model, api_key=api_key)
-
-        await ctx.report_progress(80, 100, "Processing AI-generated validation plan...")
-
-        # Get the generated validation plan
-        validation_plan = str(draft_validator)
-
-        await ctx.report_progress(90, 100, "Formatting results...")
-
-        # Extract just the Python code from the response
-        code_start = validation_plan.find("```python")
-        code_end = validation_plan.find("```", code_start + 9)
-
-        if code_start != -1 and code_end != -1:
-            python_code = validation_plan[code_start + 9 : code_end].strip()
-        else:
-            python_code = validation_plan
-
-        await ctx.report_progress(100, 100, "AI validation plan generated successfully!")
-
-        # Format the response with enhanced presentation
-        formatted_response = f"""## **🤖 AI-Generated Validation Plan**
-
-**Model Used:** {model}
-**DataFrame:** {dataframe_id} ({df.shape[0]:,} rows × {df.shape[1]} columns)
-
-### **📋 Generated Validation Code:**
-
-<details>
-<summary><strong>🔍 View Complete Python Code</strong></summary>
-
-```python
-{python_code}
-```
-</details>
-
-### **🎯 Next Steps:**
-1. **Review the generated plan** - The AI has analyzed your data structure and suggested appropriate validations
-2. **Customize as needed** - Adjust thresholds, add business-specific rules, or remove unnecessary checks
-3. **Copy and adapt** - Replace `your_data` with your actual DataFrame variable
-4. **Run the validation** - Execute the code to validate your data
-
-### **💡 Key Features:**
-- **Schema validation** ensuring column types match expectations
-- **Range checks** for numeric values based on actual data distribution
-- **Null value handling** customized to your data's missing value patterns
-- **Business logic** inferred from data characteristics
-- **Row and column count validation** ensuring data integrity
-
-The AI has analyzed your specific dataset and generated validation rules tailored to your data's characteristics!
-"""
-
+        from pointblank.draft import DraftValidation
+    except ImportError as e:
         return {
-            "status": "success",
-            "model_used": model,
-            "dataframe_id": dataframe_id,
-            "validation_plan": python_code,
-            "formatted_response": formatted_response,
-            "raw_response": validation_plan,
+            "error": "DraftValidation not available",
+            "message": "Install with: pip install pointblank[generate]",
+            "details": str(e),
         }
 
-    except Exception as e:
-        error_msg = f"Error generating AI validation plan: {str(e)}"
-        logger.error(error_msg)
-        await ctx.report_progress(100, 100, error_msg)
+    await ctx.report_progress(30, 100, "Analyzing data and generating validation plan...")
 
-        return {
-            "status": "error",
-            "error": error_msg,
-            "suggestion": "Check your API key and model availability. For Anthropic models, ensure ANTHROPIC_API_KEY is set in environment or .env file.",
-        }
+    draft_validator = DraftValidation(data=df, model=model, api_key=api_key)
+
+    await ctx.report_progress(80, 100, "Processing AI-generated validation plan...")
+
+    validation_plan = str(draft_validator)
+
+    code_start = validation_plan.find("```python")
+    code_end = validation_plan.find("```", code_start + 9)
+    python_code = (
+        validation_plan[code_start + 9 : code_end].strip()
+        if code_start != -1 and code_end != -1
+        else validation_plan
+    )
+
+    await ctx.report_progress(100, 100, "AI validation plan generated!")
+
+    return {
+        "status": "success",
+        "model_used": model,
+        "dataframe_id": dataframe_id,
+        "validation_plan": python_code,
+        "raw_response": validation_plan,
+    }
 
 
 @mcp.tool()
 def validation_assistant(
     ctx: Context,
     dataframe_id: str,
-    validation_goal: str = "general data quality",
+    validation_goal: str = "general",
 ) -> str:
     """
-    Interactive assistant to help create a validation plan for your data.
+    Generate data-aware validation suggestions based on actual column types and statistics.
 
-    This tool walks you through creating appropriate validation rules based on
-    your data characteristics and validation goals.
+    Analyzes your data's structure (types, null counts, value ranges, cardinality)
+    to produce actionable Pointblank validation code tailored to your dataset.
     """
-    try:
-        # Get the DataFrame
-        app_ctx: AppContext = ctx.request_context.lifespan_context
+    app_ctx: AppContext = ctx.request_context.lifespan_context
+    dataframe_id = validate_resource_id(dataframe_id, "DataFrame")
 
-        if dataframe_id not in app_ctx.loaded_dataframes:
-            return f"❌ Error: DataFrame '{dataframe_id}' not found. Load a DataFrame first."
+    if dataframe_id not in app_ctx.loaded_dataframes:
+        return f"Error: DataFrame '{dataframe_id}' not found. Load a DataFrame first."
 
-        data = app_ctx.loaded_dataframes[dataframe_id]
+    data = app_ctx.loaded_dataframes[dataframe_id]
+    rows, cols = data.shape
 
-        # Analyze the DataFrame structure
+    # Analyze each column with actual data characteristics
+    column_analyses = []
+    validation_steps = []
 
-        # Get basic info about the data
-        if hasattr(data, "shape"):
-            rows, cols = data.shape
-        elif hasattr(data, "count"):
-            rows = data.count().collect()[0, 0] if hasattr(data.count(), "collect") else len(data)
-            cols = len(data.columns)
+    for col in data.columns:
+        # Get dtype
+        if hasattr(data, "dtypes") and hasattr(data.dtypes, "items"):  # pandas
+            dtype = str(data.dtypes[col])
+        elif hasattr(data, "dtypes"):  # polars
+            dtype = str(data.dtypes[data.columns.index(col)])
         else:
-            rows, cols = "unknown", len(data.columns) if hasattr(data, "columns") else "unknown"
+            dtype = "unknown"
 
-        # Get column information
-        column_info = []
-        for col in data.columns:
-            if hasattr(data, "dtypes"):
-                if hasattr(data.dtypes, "items"):  # pandas
-                    dtype = str(data.dtypes[col])
-                else:  # polars
-                    dtype = str(data.dtypes[data.columns.index(col)])
+        # Count nulls
+        try:
+            if hasattr(data[col], "isnull"):  # pandas
+                null_count = int(data[col].isnull().sum())
+            elif hasattr(data[col], "null_count"):  # polars
+                null_count = int(data[col].null_count())
             else:
-                dtype = "unknown"
-            column_info.append(f"  - {col}: {dtype}")
+                null_count = 0
+        except Exception:
+            null_count = 0
 
-        # Generate validation suggestions based on goal
-        suggestions = []
+        null_pct = round(null_count / rows * 100, 1) if rows > 0 else 0
 
-        if validation_goal in ["general", "data_quality"]:
-            suggestions.extend(
-                [
-                    "# Basic Data Quality Checks",
-                    "validator = pb.Validate(data)",
-                    f".col_exists(columns={data.columns})  # Ensure all expected columns exist",
-                ]
-            )
+        col_info = {"name": col, "dtype": dtype, "null_count": null_count, "null_pct": null_pct}
 
-            # Add type-specific suggestions
-            for col in data.columns:
-                if any(keyword in col.lower() for keyword in ["id", "key"]):
-                    suggestions.append(
-                        f".col_vals_not_null(columns='{col}')  # ID columns should not be null"
-                    )
-                    suggestions.append(
-                        f".col_vals_unique(columns='{col}')  # ID columns should be unique"
-                    )
-                elif any(keyword in col.lower() for keyword in ["email", "mail"]):
-                    suggestions.append(
-                        f".col_vals_regex(columns='{col}', regex=r'[^@]+@[^@]+\\.[^@]+')  # Email format"
-                    )
-                elif any(keyword in col.lower() for keyword in ["phone", "tel"]):
-                    suggestions.append(
-                        f".col_vals_regex(columns='{col}', regex=r'\\+?[\\d\\s\\-\\(\\)]+')  # Phone format"
-                    )
-                elif any(keyword in col.lower() for keyword in ["date", "time"]):
-                    suggestions.append(
-                        f".col_vals_not_null(columns='{col}')  # Date columns should not be null"
-                    )
+        # Numeric analysis
+        is_numeric = any(t in dtype.lower() for t in ["int", "float", "decimal", "numeric"])
+        if is_numeric:
+            try:
+                if hasattr(data[col], "min"):
+                    col_info["min"] = float(data[col].min())
+                    col_info["max"] = float(data[col].max())
+            except Exception:
+                pass
 
-        if validation_goal in ["general", "completeness"]:
-            suggestions.extend(
-                [
-                    "# Completeness Checks",
-                    ".col_vals_not_null(columns=['critical_column1', 'critical_column2'])  # Critical fields must be complete",
-                    ".row_count_match(count=expected_count)  # Verify expected number of records",
-                ]
-            )
+        # Cardinality for string/categorical
+        is_string = any(
+            t in dtype.lower() for t in ["object", "string", "utf8", "str", "categorical"]
+        )
+        if is_string or (not is_numeric):
+            try:
+                if hasattr(data[col], "nunique"):
+                    col_info["n_unique"] = int(data[col].nunique())
+                elif hasattr(data[col], "n_unique"):
+                    col_info["n_unique"] = int(data[col].n_unique())
+            except Exception:
+                pass
 
-        if validation_goal in ["general", "consistency"]:
-            suggestions.extend(
-                [
-                    "# Consistency Checks",
-                    ".col_vals_in_set(columns='status', set=['active', 'inactive', 'pending'])  # Valid status values",
-                    ".col_vals_between(columns='age', left=0, right=120)  # Reasonable age range",
-                ]
-            )
+        column_analyses.append(col_info)
 
-        if validation_goal in ["general", "accuracy"]:
-            suggestions.extend(
-                [
-                    "# Accuracy Checks",
-                    ".col_vals_gt(columns='price', value=0)  # Prices should be positive",
-                    ".col_vals_le(columns='discount_pct', value=100)  # Discounts <= 100%",
-                ]
-            )
+        # Generate validation suggestions based on actual data characteristics
+        if null_count == 0:
+            validation_steps.append(f"    .col_vals_not_null(columns='{col}')  # Currently 0% null")
+        elif null_pct > 50:
+            # Don't validate not-null for mostly-null columns
+            pass
 
-        # Add interrogation
-        suggestions.append(".interrogate()  # Execute all validation steps")
+        if is_numeric and "min" in col_info and "max" in col_info:
+            min_val = col_info["min"]
+            max_val = col_info["max"]
+            if min_val >= 0:
+                validation_steps.append(
+                    f"    .col_vals_ge(columns='{col}', value=0)  # Range: [{min_val}, {max_val}]"
+                )
+            else:
+                validation_steps.append(
+                    f"    .col_vals_between(columns='{col}', left={min_val}, right={max_val})"
+                )
 
-        suggestion_text = "\n".join(suggestions)
+        if is_string and "n_unique" in col_info:
+            n_unique = col_info["n_unique"]
+            if n_unique <= 10 and rows > 20:
+                # Low cardinality - suggest in_set
+                try:
+                    if hasattr(data[col], "dropna"):  # pandas
+                        unique_vals = list(data[col].dropna().unique()[:10])
+                    elif hasattr(data[col], "drop_nulls"):  # polars
+                        unique_vals = data[col].drop_nulls().unique().to_list()[:10]
+                    else:
+                        unique_vals = None
 
-        response = f"""
-🔍 **Validation Assistant for DataFrame '{dataframe_id}'**
+                    if unique_vals:
+                        validation_steps.append(
+                            f"    .col_vals_in_set(columns='{col}', set={repr(unique_vals)})"
+                        )
+                except Exception:
+                    pass
 
-📊 **Data Overview:**
-- Rows: {rows}
-- Columns: {cols}
-- Column Details:
-{chr(10).join(column_info)}
+    # Build column details section
+    col_details = []
+    for ca in column_analyses:
+        detail = f"  - **{ca['name']}** ({ca['dtype']}): {ca['null_pct']}% null"
+        if "min" in ca:
+            detail += f", range [{ca['min']}, {ca['max']}]"
+        if "n_unique" in ca:
+            detail += f", {ca['n_unique']} unique values"
+        col_details.append(detail)
 
-🎯 **Validation Goal:** {validation_goal}
+    # Build validation code
+    code = "import pointblank as pb\n\nvalidator = (\n    pb.Validate(data)\n"
+    if validation_steps:
+        code += "\n".join(validation_steps)
+    else:
+        code += "    # No automatic suggestions - add your own validation steps"
+    code += "\n    .interrogate()\n)"
 
-📋 **Suggested Validation Plan:**
+    response = f"""**Validation Assistant for '{dataframe_id}'**
+
+**Data Overview:** {rows:,} rows x {cols} columns
+
+**Column Analysis:**
+{chr(10).join(col_details)}
+
+**Suggested Validation Plan ({len(validation_steps)} rules):**
 
 ```python
-{suggestion_text}
+{code}
 ```
 
-💡 **Next Steps:**
-1. Review the suggested validation rules above
-2. Customize the rules based on your specific business requirements
-3. Use the `create_validator` tool to implement these checks
-4. Run `interrogate_validator` to execute the validation
-
-❓ **Need More Help?**
-- Use `validation_goal='data_quality'` for basic data quality checks
-- Use `validation_goal='completeness'` for missing data validation
-- Use `validation_goal='consistency'` for value range/format validation
-- Use `validation_goal='accuracy'` for business rule validation
-
-Would you like me to create a validator with these suggestions? Use the `create_validator` tool with the above validation steps!
+**Next Steps:**
+1. Review and customize the suggestions above
+2. Use `create_validator` + `add_validation_step` to implement
+3. Or use `draft_validation_plan` for AI-powered suggestions
 """
 
-        return response
+    return response
 
-    except Exception as e:
-        logger.error(f"Error in validation assistant: {e}")
-        return f"❌ Error in validation assistant: {str(e)}"
+
+# =============================================================================
+# TOOLS: Server Management
+# =============================================================================
+
+
+@mcp.tool(
+    name="server_health_check",
+    description="Get comprehensive server health and status information.",
+    tags={"Server Management"},
+)
+async def server_health_check(ctx: Context) -> Dict[str, Any]:
+    """Returns server health information."""
+    app_ctx: AppContext = ctx.request_context.lifespan_context
+
+    dataframes_count = len(app_ctx.loaded_dataframes)
+    validators_count = len(app_ctx.active_validators)
+
+    total_memory_mb = 0
+    for df in app_ctx.loaded_dataframes.values():
+        try:
+            if hasattr(df, "memory_usage"):
+                total_memory_mb += df.memory_usage(deep=True).sum() / 1024 / 1024
+            else:
+                total_memory_mb += df.shape[0] * df.shape[1] * 8 / 1024 / 1024
+        except Exception:
+            pass
+
+    health_info = {
+        "timestamp": datetime.now().isoformat(),
+        "server_status": "healthy",
+        "system_info": {
+            "python_version": sys.version.split()[0],
+            "platform": platform.platform(),
+        },
+        "backend_status": {
+            "pandas_available": HAS_PANDAS,
+            "polars_available": HAS_POLARS,
+        },
+        "resource_usage": {
+            "total_dataframes": dataframes_count,
+            "total_validators": validators_count,
+            "total_memory_mb": round(total_memory_mb, 2),
+        },
+        "capabilities": {
+            "supported_file_formats": ["CSV", "JSON", "JSONL", "Parquet"]
+            + (["Excel"] if HAS_PANDAS else []),
+            "validation_types_count": len(SUPPORTED_VALIDATION_TYPES),
+            "templates_available": AVAILABLE_TEMPLATES,
+        },
+    }
+
+    warnings = []
+    if not HAS_PANDAS and not HAS_POLARS:
+        warnings.append("No DataFrame backends available")
+        health_info["server_status"] = "degraded"
+    if total_memory_mb > 1000:
+        warnings.append(f"High memory usage: {total_memory_mb:.1f}MB")
+    if warnings:
+        health_info["warnings"] = warnings
+
+    return health_info
 
 
 @mcp.tool(
@@ -2327,214 +1142,79 @@ async def get_pointblank_api_reference(
     ctx: Context,
     category: Annotated[
         str,
-        "Category of API reference: 'validation_methods', 'data_types', 'thresholds', 'common_patterns', or 'all'",
+        "Category: 'validation_methods', 'thresholds', 'common_patterns', or 'all'",
     ] = "validation_methods",
 ) -> str:
-    """
-    Provides comprehensive API reference for Pointblank validation methods and patterns.
-    Helps ensure correct parameter usage and method signatures.
-    """
+    """Provides API reference for Pointblank validation methods and patterns."""
 
-    validation_methods = {
-        "col_vals_ge": {
-            "description": "Check that column values are greater than or equal to value",
-            "parameters": {
-                "columns": "str | list - Column name(s) to check",
-                "value": "float - Minimum value",
-            },
-            "example": '.col_vals_ge(columns="price", value=0)',
-        },
-        "col_vals_gt": {
-            "description": "Check that column values are greater than value",
-            "parameters": {
-                "columns": "str | list - Column name(s) to check",
-                "value": "float - Value to compare against",
-            },
-            "example": '.col_vals_gt(columns="quantity", value=0)',
-        },
-        "col_vals_le": {
-            "description": "Check that column values are less than or equal to value",
-            "parameters": {
-                "columns": "str | list - Column name(s) to check",
-                "value": "float - Maximum value",
-            },
-            "example": '.col_vals_le(columns="percentage", value=100)',
-        },
-        "col_vals_lt": {
-            "description": "Check that column values are less than value",
-            "parameters": {
-                "columns": "str | list - Column name(s) to check",
-                "value": "float - Value to compare against",
-            },
-            "example": '.col_vals_lt(columns="score", value=100)',
-        },
-        "col_vals_null": {
-            "description": "Check that column values are null/missing",
-            "parameters": {"columns": "str | list - Column name(s) to check"},
-            "example": '.col_vals_null(columns="empty")',
-        },
-        "col_vals_not_null": {
-            "description": "Check that column values are not null/missing",
-            "parameters": {"columns": "str | list - Column name(s) to check"},
-            "example": '.col_vals_not_null(columns="email")',
-        },
-        "col_vals_between": {
-            "description": "Check that column values are within a numeric range",
-            "parameters": {
-                "columns": "str | list - Column name(s) to check",
-                "left": "float - Lower bound (inclusive by default)",
-                "right": "float - Upper bound (inclusive by default)",
-                "inclusive": "tuple[bool, bool] - (left_inclusive, right_inclusive)",
-            },
-            "example": '.col_vals_between(columns="age", left=0, right=120)',
-        },
-        "col_vals_in_set": {
-            "description": "Check that column values are in allowed set",
-            "parameters": {
-                "columns": "str | list - Column name(s) to check",
-                "set": "list - List of allowed values",
-            },
-            "example": '.col_vals_in_set(columns="status", set=["active", "inactive"])',
-        },
-        "col_vals_outside": {
-            "description": "Check that column values are outside of a specified set",
-            "parameters": {
-                "columns": "str | list - Column name(s) to check",
-                "set": "list - List of allowed values",
-            },
-            "example": '.col_vals_outside(columns="status", set=["active", "inactive"])',
-        },
-        "col_vals_regex": {
-            "description": "Check that column values match regex pattern",
-            "parameters": {
-                "columns": "str | list - Column name(s) to check",
-                "pattern": "str - Regular expression pattern",
-            },
-            "example": '.col_vals_regex(columns="email", pattern=r"[^@]+@[^@]+\\.[^@]+")',
-        },
-        "col_exists": {
-            "description": "Check that specified columns exist in the table",
-            "parameters": {"columns": "str | list - Column name(s) to check"},
-            "example": '.col_exists(columns=["name", "email", "age"])',
-        },
-        "rows_distinct": {
-            "description": "Check that all rows in the table are unique",
-            "parameters": {
-                "columns_subset": "str | list - Column name(s) for constraining uniqueness"
-            },
-            "example": ".rows_distinct()",
-        },
-        "rows_complete": {
-            "description": "Check that all rows in the table are complete (no missing values)",
-            "parameters": {
-                "columns_subset": "str | list - Column name(s) for constraining completeness"
-            },
-            "example": ".rows_complete()",
-        },
-    }
+    validation_methods_ref = """# Pointblank Validation Methods
 
-    data_types_info = """
-    **Column Data Type Expectations:**
-    - Numeric validations (between, gt, ge, lt, le): Work with int, float columns
-    - String validations (regex, in_set): Work with object/string columns
-    - Null checks (not_null, null): Work with any column type
-    - Set validations (in_set, not_in_set): Work with any comparable type
-    """
+## Value Comparisons
+- `col_vals_gt(columns, value)` - Greater than
+- `col_vals_ge(columns, value)` - Greater than or equal
+- `col_vals_lt(columns, value)` - Less than
+- `col_vals_le(columns, value)` - Less than or equal
+- `col_vals_eq(columns, value)` - Equal to
+- `col_vals_ne(columns, value)` - Not equal to
+- `col_vals_between(columns, left, right)` - Within range (inclusive)
+- `col_vals_outside(columns, left, right)` - Outside range
 
-    thresholds_info = """
-    **Threshold Configuration:**
-    ```python
-    thresholds = {
-        "warning": 0.05,    # 5% failures trigger warning
-        "error": 0.10,      # 10% failures trigger error
-        "critical": 0.15    # 15% failures trigger critical
-    }
-    ```
+## Set Membership
+- `col_vals_in_set(columns, set)` - Values in allowed set
+- `col_vals_not_in_set(columns, set)` - Values not in set
 
-    **Threshold Levels:**
-    - warning: Minor data quality issues
-    - error: Significant problems requiring attention
-    - critical: Severe issues that stop processing
-    """
+## Null Checks
+- `col_vals_null(columns)` - Values are null
+- `col_vals_not_null(columns)` - Values are not null
 
-    common_patterns = """
-    **Common Validation Patterns:**
+## Pattern Matching
+- `col_vals_regex(columns, pattern)` - Match regex pattern
 
-    1. **Data Integrity:**
-    ```python
-    .col_vals_not_null(columns="id")
-    .rows_distinct()
-    .col_exists(columns=["required_field1", "required_field2"])
-    ```
-
-    2. **Business Rules:**
-    ```python
-    .col_vals_between(columns="age", left=0, right=120)
-    .col_vals_in_set(columns="status", set=["active", "inactive", "pending"])
-    .col_vals_ge(columns="price", value=0)
-    ```
-
-    3. **Format Validation:**
-    ```python
-    .col_vals_regex(columns="email", pattern=r"[^@]+@[^@]+\\.[^@]+")
-    .col_vals_regex(columns="phone", pattern=r"\\d{3}-\\d{3}-\\d{4}")
-    ```
-
-    4. **Range Validation:**
-    ```python
-    .col_vals_between(columns="percentage", left=0, right=100)
-    .col_vals_between(columns="year", left=1900, right=2030)
-    ```
-    """
-
-    if category == "validation_methods":
-        result = "# 🔧 **Pointblank Validation Methods Reference**\n\n"
-        for method_name, info in validation_methods.items():
-            result += f"## `{method_name}`\n"
-            result += f"**Description:** {info['description']}\n\n"
-            result += "**Parameters:**\n"
-            for param, desc in info["parameters"].items():
-                result += f"- `{param}`: {desc}\n"
-            result += f"\n**Example:** `{info['example']}`\n\n---\n\n"
-        return result
-
-    elif category == "data_types":
-        return f"# 📊 **Data Types Reference**\n\n{data_types_info}"
-
-    elif category == "thresholds":
-        return f"# ⚠️ **Thresholds Reference**\n\n{thresholds_info}"
-
-    elif category == "common_patterns":
-        return f"# 🎯 **Common Patterns Reference**\n\n{common_patterns}"
-
-    elif category == "all":
-        # Build complete reference
-        validation_methods_ref = "# 🔧 **Pointblank Validation Methods Reference**\n\n"
-        for method_name, info in validation_methods.items():
-            validation_methods_ref += f"## `{method_name}`\n"
-            validation_methods_ref += f"**Description:** {info['description']}\n\n"
-            validation_methods_ref += "**Parameters:**\n"
-            for param, desc in info["parameters"].items():
-                validation_methods_ref += f"- `{param}`: {desc}\n"
-            validation_methods_ref += f"\n**Example:** `{info['example']}`\n\n---\n\n"
-
-        return f"""# 📚 **Complete Pointblank API Reference**
-
-{validation_methods_ref}
-
-{data_types_info}
-
-{thresholds_info}
-
-{common_patterns}
-
-**💡 Tips:**
-- Always check parameter names match exactly (e.g., 'columns' not 'column')
-- Use list for multiple columns: `columns=["col1", "col2"]`
-- String values in sets need quotes: `set=["value1", "value2"]`
-- Numeric ranges are inclusive by default
-- Regular expressions need proper escaping
+## Table Checks
+- `col_exists(columns)` - Column exists
+- `rows_distinct()` - All rows unique
+- `rows_complete()` - No nulls in specified columns
+- `row_count_match(count)` - Expected row count
+- `col_count_match(count)` - Expected column count
+- `col_schema_match(schema)` - Schema validation
 """
 
+    thresholds_ref = """# Thresholds
+
+```python
+thresholds = {"warning": 0.05, "error": 0.10, "critical": 0.15}
+```
+
+- **warning**: Minor issues (e.g., 5% failures)
+- **error**: Significant problems (e.g., 10% failures)
+- **critical**: Severe issues that stop processing
+"""
+
+    patterns_ref = """# Common Patterns
+
+```python
+# Data Integrity
+.col_vals_not_null(columns="id")
+.rows_distinct()
+.col_exists(columns=["required_field1", "required_field2"])
+
+# Business Rules
+.col_vals_between(columns="age", left=0, right=120)
+.col_vals_in_set(columns="status", set=["active", "inactive"])
+.col_vals_ge(columns="price", value=0)
+
+# Format Validation
+.col_vals_regex(columns="email", pattern=r"[^@]+@[^@]+\\.[^@]+")
+```
+"""
+
+    if category == "validation_methods":
+        return validation_methods_ref
+    elif category == "thresholds":
+        return thresholds_ref
+    elif category == "common_patterns":
+        return patterns_ref
+    elif category == "all":
+        return f"{validation_methods_ref}\n{thresholds_ref}\n{patterns_ref}"
     else:
-        return f"❌ Unknown category '{category}'. Use: validation_methods, data_types, thresholds, common_patterns, or all"
+        return f"Unknown category '{category}'. Use: validation_methods, thresholds, common_patterns, or all"
